Crate rusoto_ec2

Amazon Elastic Compute Cloud
Amazon Elastic Compute Cloud (Amazon EC2) provides resizable computing capacity in the Amazon Web Services (AWS) cloud. Using Amazon EC2 eliminates your need to invest in hardware up front, so you can develop and deploy applications faster.

If you're using the service, you're probably looking for Ec2Client and Ec2.

Structs

AcceptReservedInstancesExchangeQuoteRequest
Contains the parameters for accepting the quote.

AcceptReservedInstancesExchangeQuoteResult
The result of the exchange and whether it was successful.

AcceptVpcPeeringConnectionRequest
Contains the parameters for AcceptVpcPeeringConnection.

AcceptVpcPeeringConnectionResult
Contains the output of AcceptVpcPeeringConnection.

AccountAttribute
Describes an account attribute.

AccountAttributeValue
Describes a value of an account attribute.

ActiveInstance
Describes a running instance in a Spot fleet.

Address
Describes an Elastic IP address.

AllocateAddressRequest
Contains the parameters for AllocateAddress.

AllocateAddressResult
Contains the output of AllocateAddress.

AllocateHostsRequest
Contains the parameters for AllocateHosts.

AllocateHostsResult
Contains the output of AllocateHosts.

AssignIpv6AddressesRequest

AssignIpv6AddressesResult

AssignPrivateIpAddressesRequest
Contains the parameters for AssignPrivateIpAddresses.

AssociateAddressRequest
Contains the parameters for AssociateAddress.

AssociateAddressResult
Contains the output of AssociateAddress.

AssociateDhcpOptionsRequest
Contains the parameters for AssociateDhcpOptions.

AssociateIamInstanceProfileRequest

AssociateIamInstanceProfileResult

AssociateRouteTableRequest
Contains the parameters for AssociateRouteTable.

AssociateRouteTableResult
Contains the output of AssociateRouteTable.

AssociateSubnetCidrBlockRequest

AssociateSubnetCidrBlockResult

AssociateVpcCidrBlockRequest

AssociateVpcCidrBlockResult

AttachClassicLinkVpcRequest
Contains the parameters for AttachClassicLinkVpc.

AttachClassicLinkVpcResult
Contains the output of AttachClassicLinkVpc.

AttachInternetGatewayRequest
Contains the parameters for AttachInternetGateway.

AttachNetworkInterfaceRequest
Contains the parameters for AttachNetworkInterface.

AttachNetworkInterfaceResult
Contains the output of AttachNetworkInterface.

AttachVolumeRequest
Contains the parameters for AttachVolume.

AttachVpnGatewayRequest
Contains the parameters for AttachVpnGateway.

AttachVpnGatewayResult
Contains the output of AttachVpnGateway.

AttributeBooleanValue
Describes a value for a resource attribute that is a Boolean value.

AttributeValue
Describes a value for a resource attribute that is a String.

AuthorizeSecurityGroupEgressRequest
Contains the parameters for AuthorizeSecurityGroupEgress.

AuthorizeSecurityGroupIngressRequest
Contains the parameters for AuthorizeSecurityGroupIngress.

AvailabilityZone
Describes an Availability Zone.

AvailabilityZoneMessage
Describes a message about an Availability Zone.

AvailableCapacity
The capacity information for instances launched onto the Dedicated Host.

BlobAttributeValue

BlockDeviceMapping
Describes a block device mapping.

BundleInstanceRequest
Contains the parameters for BundleInstance.

BundleInstanceResult
Contains the output of BundleInstance.

BundleTask
Describes a bundle task.

BundleTaskError
Describes an error for BundleInstance.

CancelBundleTaskRequest
Contains the parameters for CancelBundleTask.

CancelBundleTaskResult
Contains the output of CancelBundleTask.

CancelConversionRequest
Contains the parameters for CancelConversionTask.

CancelExportTaskRequest
Contains the parameters for CancelExportTask.

CancelImportTaskRequest
Contains the parameters for CancelImportTask.

CancelImportTaskResult
Contains the output for CancelImportTask.

CancelReservedInstancesListingRequest
Contains the parameters for CancelReservedInstancesListing.

CancelReservedInstancesListingResult
Contains the output of CancelReservedInstancesListing.

CancelSpotFleetRequestsError
Describes a Spot fleet error.

CancelSpotFleetRequestsErrorItem
Describes a Spot fleet request that was not successfully canceled.

CancelSpotFleetRequestsRequest
Contains the parameters for CancelSpotFleetRequests.

CancelSpotFleetRequestsResponse
Contains the output of CancelSpotFleetRequests.

CancelSpotFleetRequestsSuccessItem
Describes a Spot fleet request that was successfully canceled.

CancelSpotInstanceRequestsRequest
Contains the parameters for CancelSpotInstanceRequests.

CancelSpotInstanceRequestsResult
Contains the output of CancelSpotInstanceRequests.

CancelledSpotInstanceRequest
Describes a request to cancel a Spot instance.

ClassicLinkDnsSupport
Describes the ClassicLink DNS support status of a VPC.

ClassicLinkInstance
Describes a linked EC2-Classic instance.

ClientData
Describes the client-specific data.

ConfirmProductInstanceRequest
Contains the parameters for ConfirmProductInstance.

ConfirmProductInstanceResult
Contains the output of ConfirmProductInstance.

ConversionTask
Describes a conversion task.

CopyImageRequest
Contains the parameters for CopyImage.

CopyImageResult
Contains the output of CopyImage.

CopySnapshotRequest
Contains the parameters for CopySnapshot.

CopySnapshotResult
Contains the output of CopySnapshot.

CreateCustomerGatewayRequest
Contains the parameters for CreateCustomerGateway.

CreateCustomerGatewayResult
Contains the output of CreateCustomerGateway.

CreateDhcpOptionsRequest
Contains the parameters for CreateDhcpOptions.

CreateDhcpOptionsResult
Contains the output of CreateDhcpOptions.

CreateEgressOnlyInternetGatewayRequest

CreateEgressOnlyInternetGatewayResult

CreateFlowLogsRequest
Contains the parameters for CreateFlowLogs.

CreateFlowLogsResult
Contains the output of CreateFlowLogs.

CreateFpgaImageRequest

CreateFpgaImageResult

CreateImageRequest
Contains the parameters for CreateImage.

CreateImageResult
Contains the output of CreateImage.

CreateInstanceExportTaskRequest
Contains the parameters for CreateInstanceExportTask.

CreateInstanceExportTaskResult
Contains the output for CreateInstanceExportTask.

CreateInternetGatewayRequest
Contains the parameters for CreateInternetGateway.

CreateInternetGatewayResult
Contains the output of CreateInternetGateway.

CreateKeyPairRequest
Contains the parameters for CreateKeyPair.

CreateNatGatewayRequest
Contains the parameters for CreateNatGateway.

CreateNatGatewayResult
Contains the output of CreateNatGateway.

CreateNetworkAclEntryRequest
Contains the parameters for CreateNetworkAclEntry.

CreateNetworkAclRequest
Contains the parameters for CreateNetworkAcl.

CreateNetworkAclResult
Contains the output of CreateNetworkAcl.

CreateNetworkInterfaceRequest
Contains the parameters for CreateNetworkInterface.

CreateNetworkInterfaceResult
Contains the output of CreateNetworkInterface.

CreatePlacementGroupRequest
Contains the parameters for CreatePlacementGroup.

CreateReservedInstancesListingRequest
Contains the parameters for CreateReservedInstancesListing.

CreateReservedInstancesListingResult
Contains the output of CreateReservedInstancesListing.

CreateRouteRequest
Contains the parameters for CreateRoute.

CreateRouteResult
Contains the output of CreateRoute.

CreateRouteTableRequest
Contains the parameters for CreateRouteTable.

CreateRouteTableResult
Contains the output of CreateRouteTable.

CreateSecurityGroupRequest
Contains the parameters for CreateSecurityGroup.

CreateSecurityGroupResult
Contains the output of CreateSecurityGroup.

CreateSnapshotRequest
Contains the parameters for CreateSnapshot.

CreateSpotDatafeedSubscriptionRequest
Contains the parameters for CreateSpotDatafeedSubscription.

CreateSpotDatafeedSubscriptionResult
Contains the output of CreateSpotDatafeedSubscription.

CreateSubnetRequest
Contains the parameters for CreateSubnet.

CreateSubnetResult
Contains the output of CreateSubnet.

CreateTagsRequest
Contains the parameters for CreateTags.

CreateVolumePermission
Describes the user or group to be added or removed from the permissions for a volume.

CreateVolumePermissionModifications
Describes modifications to the permissions for a volume.

CreateVolumeRequest
Contains the parameters for CreateVolume.

CreateVpcEndpointRequest
Contains the parameters for CreateVpcEndpoint.

CreateVpcEndpointResult
Contains the output of CreateVpcEndpoint.

CreateVpcPeeringConnectionRequest
Contains the parameters for CreateVpcPeeringConnection.

CreateVpcPeeringConnectionResult
Contains the output of CreateVpcPeeringConnection.

CreateVpcRequest
Contains the parameters for CreateVpc.

CreateVpcResult
Contains the output of CreateVpc.

CreateVpnConnectionRequest
Contains the parameters for CreateVpnConnection.

CreateVpnConnectionResult
Contains the output of CreateVpnConnection.

CreateVpnConnectionRouteRequest
Contains the parameters for CreateVpnConnectionRoute.

CreateVpnGatewayRequest
Contains the parameters for CreateVpnGateway.

CreateVpnGatewayResult
Contains the output of CreateVpnGateway.

CustomerGateway
Describes a customer gateway.

DeleteCustomerGatewayRequest
Contains the parameters for DeleteCustomerGateway.

DeleteDhcpOptionsRequest
Contains the parameters for DeleteDhcpOptions.

DeleteEgressOnlyInternetGatewayRequest

DeleteEgressOnlyInternetGatewayResult

DeleteFlowLogsRequest
Contains the parameters for DeleteFlowLogs.

DeleteFlowLogsResult
Contains the output of DeleteFlowLogs.

DeleteInternetGatewayRequest
Contains the parameters for DeleteInternetGateway.

DeleteKeyPairRequest
Contains the parameters for DeleteKeyPair.

DeleteNatGatewayRequest
Contains the parameters for DeleteNatGateway.

DeleteNatGatewayResult
Contains the output of DeleteNatGateway.

DeleteNetworkAclEntryRequest
Contains the parameters for DeleteNetworkAclEntry.

DeleteNetworkAclRequest
Contains the parameters for DeleteNetworkAcl.

DeleteNetworkInterfaceRequest
Contains the parameters for DeleteNetworkInterface.

DeletePlacementGroupRequest
Contains the parameters for DeletePlacementGroup.

DeleteRouteRequest
Contains the parameters for DeleteRoute.

DeleteRouteTableRequest
Contains the parameters for DeleteRouteTable.

DeleteSecurityGroupRequest
Contains the parameters for DeleteSecurityGroup.

DeleteSnapshotRequest
Contains the parameters for DeleteSnapshot.

DeleteSpotDatafeedSubscriptionRequest
Contains the parameters for DeleteSpotDatafeedSubscription.

DeleteSubnetRequest
Contains the parameters for DeleteSubnet.

DeleteTagsRequest
Contains the parameters for DeleteTags.

DeleteVolumeRequest
Contains the parameters for DeleteVolume.

DeleteVpcEndpointsRequest
Contains the parameters for DeleteVpcEndpoints.

DeleteVpcEndpointsResult
Contains the output of DeleteVpcEndpoints.

DeleteVpcPeeringConnectionRequest
Contains the parameters for DeleteVpcPeeringConnection.

DeleteVpcPeeringConnectionResult
Contains the output of DeleteVpcPeeringConnection.

DeleteVpcRequest
Contains the parameters for DeleteVpc.

DeleteVpnConnectionRequest
Contains the parameters for DeleteVpnConnection.

DeleteVpnConnectionRouteRequest
Contains the parameters for DeleteVpnConnectionRoute.

DeleteVpnGatewayRequest
Contains the parameters for DeleteVpnGateway.

DeregisterImageRequest
Contains the parameters for DeregisterImage.

DescribeAccountAttributesRequest
Contains the parameters for DescribeAccountAttributes.

DescribeAccountAttributesResult
Contains the output of DescribeAccountAttributes.

DescribeAddressesRequest
Contains the parameters for DescribeAddresses.

DescribeAddressesResult
Contains the output of DescribeAddresses.

DescribeAvailabilityZonesRequest
Contains the parameters for DescribeAvailabilityZones.

DescribeAvailabilityZonesResult
Contains the output of DescribeAvailabiltyZones.

DescribeBundleTasksRequest
Contains the parameters for DescribeBundleTasks.

DescribeBundleTasksResult
Contains the output of DescribeBundleTasks.

DescribeClassicLinkInstancesRequest
Contains the parameters for DescribeClassicLinkInstances.

DescribeClassicLinkInstancesResult
Contains the output of DescribeClassicLinkInstances.

DescribeConversionTasksRequest
Contains the parameters for DescribeConversionTasks.

DescribeConversionTasksResult
Contains the output for DescribeConversionTasks.

DescribeCustomerGatewaysRequest
Contains the parameters for DescribeCustomerGateways.

DescribeCustomerGatewaysResult
Contains the output of DescribeCustomerGateways.

DescribeDhcpOptionsRequest
Contains the parameters for DescribeDhcpOptions.

DescribeDhcpOptionsResult
Contains the output of DescribeDhcpOptions.

DescribeEgressOnlyInternetGatewaysRequest

DescribeEgressOnlyInternetGatewaysResult

DescribeExportTasksRequest
Contains the parameters for DescribeExportTasks.

DescribeExportTasksResult
Contains the output for DescribeExportTasks.

DescribeFlowLogsRequest
Contains the parameters for DescribeFlowLogs.

DescribeFlowLogsResult
Contains the output of DescribeFlowLogs.

DescribeFpgaImagesRequest

DescribeFpgaImagesResult

DescribeHostReservationOfferingsRequest

DescribeHostReservationOfferingsResult

DescribeHostReservationsRequest

DescribeHostReservationsResult

DescribeHostsRequest
Contains the parameters for DescribeHosts.

DescribeHostsResult
Contains the output of DescribeHosts.

DescribeIamInstanceProfileAssociationsRequest

DescribeIamInstanceProfileAssociationsResult

DescribeIdFormatRequest
Contains the parameters for DescribeIdFormat.

DescribeIdFormatResult
Contains the output of DescribeIdFormat.

DescribeIdentityIdFormatRequest
Contains the parameters for DescribeIdentityIdFormat.

DescribeIdentityIdFormatResult
Contains the output of DescribeIdentityIdFormat.

DescribeImageAttributeRequest
Contains the parameters for DescribeImageAttribute.

DescribeImagesRequest
Contains the parameters for DescribeImages.

DescribeImagesResult
Contains the output of DescribeImages.

DescribeImportImageTasksRequest
Contains the parameters for DescribeImportImageTasks.

DescribeImportImageTasksResult
Contains the output for DescribeImportImageTasks.

DescribeImportSnapshotTasksRequest
Contains the parameters for DescribeImportSnapshotTasks.

DescribeImportSnapshotTasksResult
Contains the output for DescribeImportSnapshotTasks.

DescribeInstanceAttributeRequest
Contains the parameters for DescribeInstanceAttribute.

DescribeInstanceStatusRequest
Contains the parameters for DescribeInstanceStatus.

DescribeInstanceStatusResult
Contains the output of DescribeInstanceStatus.

DescribeInstancesRequest
Contains the parameters for DescribeInstances.

DescribeInstancesResult
Contains the output of DescribeInstances.

DescribeInternetGatewaysRequest
Contains the parameters for DescribeInternetGateways.

DescribeInternetGatewaysResult
Contains the output of DescribeInternetGateways.

DescribeKeyPairsRequest
Contains the parameters for DescribeKeyPairs.

DescribeKeyPairsResult
Contains the output of DescribeKeyPairs.

DescribeMovingAddressesRequest
Contains the parameters for DescribeMovingAddresses.

DescribeMovingAddressesResult
Contains the output of DescribeMovingAddresses.

DescribeNatGatewaysRequest
Contains the parameters for DescribeNatGateways.

DescribeNatGatewaysResult
Contains the output of DescribeNatGateways.

DescribeNetworkAclsRequest
Contains the parameters for DescribeNetworkAcls.

DescribeNetworkAclsResult
Contains the output of DescribeNetworkAcls.

DescribeNetworkInterfaceAttributeRequest
Contains the parameters for DescribeNetworkInterfaceAttribute.

DescribeNetworkInterfaceAttributeResult
Contains the output of DescribeNetworkInterfaceAttribute.

DescribeNetworkInterfacesRequest
Contains the parameters for DescribeNetworkInterfaces.

DescribeNetworkInterfacesResult
Contains the output of DescribeNetworkInterfaces.

DescribePlacementGroupsRequest
Contains the parameters for DescribePlacementGroups.

DescribePlacementGroupsResult
Contains the output of DescribePlacementGroups.

DescribePrefixListsRequest
Contains the parameters for DescribePrefixLists.

DescribePrefixListsResult
Contains the output of DescribePrefixLists.

DescribeRegionsRequest
Contains the parameters for DescribeRegions.

DescribeRegionsResult
Contains the output of DescribeRegions.

DescribeReservedInstancesListingsRequest
Contains the parameters for DescribeReservedInstancesListings.

DescribeReservedInstancesListingsResult
Contains the output of DescribeReservedInstancesListings.

DescribeReservedInstancesModificationsRequest
Contains the parameters for DescribeReservedInstancesModifications.

DescribeReservedInstancesModificationsResult
Contains the output of DescribeReservedInstancesModifications.

DescribeReservedInstancesOfferingsRequest
Contains the parameters for DescribeReservedInstancesOfferings.

DescribeReservedInstancesOfferingsResult
Contains the output of DescribeReservedInstancesOfferings.

DescribeReservedInstancesRequest
Contains the parameters for DescribeReservedInstances.

DescribeReservedInstancesResult
Contains the output for DescribeReservedInstances.

DescribeRouteTablesRequest
Contains the parameters for DescribeRouteTables.

DescribeRouteTablesResult
Contains the output of DescribeRouteTables.

DescribeScheduledInstanceAvailabilityRequest
Contains the parameters for DescribeScheduledInstanceAvailability.

DescribeScheduledInstanceAvailabilityResult
Contains the output of DescribeScheduledInstanceAvailability.

DescribeScheduledInstancesRequest
Contains the parameters for DescribeScheduledInstances.

DescribeScheduledInstancesResult
Contains the output of DescribeScheduledInstances.

DescribeSecurityGroupReferencesRequest

DescribeSecurityGroupReferencesResult

DescribeSecurityGroupsRequest
Contains the parameters for DescribeSecurityGroups.

DescribeSecurityGroupsResult
Contains the output of DescribeSecurityGroups.

DescribeSnapshotAttributeRequest
Contains the parameters for DescribeSnapshotAttribute.

DescribeSnapshotAttributeResult
Contains the output of DescribeSnapshotAttribute.

DescribeSnapshotsRequest
Contains the parameters for DescribeSnapshots.

DescribeSnapshotsResult
Contains the output of DescribeSnapshots.

DescribeSpotDatafeedSubscriptionRequest
Contains the parameters for DescribeSpotDatafeedSubscription.

DescribeSpotDatafeedSubscriptionResult
Contains the output of DescribeSpotDatafeedSubscription.

DescribeSpotFleetInstancesRequest
Contains the parameters for DescribeSpotFleetInstances.

DescribeSpotFleetInstancesResponse
Contains the output of DescribeSpotFleetInstances.

DescribeSpotFleetRequestHistoryRequest
Contains the parameters for DescribeSpotFleetRequestHistory.

DescribeSpotFleetRequestHistoryResponse
Contains the output of DescribeSpotFleetRequestHistory.

DescribeSpotFleetRequestsRequest
Contains the parameters for DescribeSpotFleetRequests.

DescribeSpotFleetRequestsResponse
Contains the output of DescribeSpotFleetRequests.

DescribeSpotInstanceRequestsRequest
Contains the parameters for DescribeSpotInstanceRequests.

DescribeSpotInstanceRequestsResult
Contains the output of DescribeSpotInstanceRequests.

DescribeSpotPriceHistoryRequest
Contains the parameters for DescribeSpotPriceHistory.

DescribeSpotPriceHistoryResult
Contains the output of DescribeSpotPriceHistory.

DescribeStaleSecurityGroupsRequest

DescribeStaleSecurityGroupsResult

DescribeSubnetsRequest
Contains the parameters for DescribeSubnets.

DescribeSubnetsResult
Contains the output of DescribeSubnets.

DescribeTagsRequest
Contains the parameters for DescribeTags.

DescribeTagsResult
Contains the output of DescribeTags.

DescribeVolumeAttributeRequest
Contains the parameters for DescribeVolumeAttribute.

DescribeVolumeAttributeResult
Contains the output of DescribeVolumeAttribute.

DescribeVolumeStatusRequest
Contains the parameters for DescribeVolumeStatus.

DescribeVolumeStatusResult
Contains the output of DescribeVolumeStatus.

DescribeVolumesModificationsRequest

DescribeVolumesModificationsResult

DescribeVolumesRequest
Contains the parameters for DescribeVolumes.

DescribeVolumesResult
Contains the output of DescribeVolumes.

DescribeVpcAttributeRequest
Contains the parameters for DescribeVpcAttribute.

DescribeVpcAttributeResult
Contains the output of DescribeVpcAttribute.

DescribeVpcClassicLinkDnsSupportRequest
Contains the parameters for DescribeVpcClassicLinkDnsSupport.

DescribeVpcClassicLinkDnsSupportResult
Contains the output of DescribeVpcClassicLinkDnsSupport.

DescribeVpcClassicLinkRequest
Contains the parameters for DescribeVpcClassicLink.

DescribeVpcClassicLinkResult
Contains the output of DescribeVpcClassicLink.

DescribeVpcEndpointServicesRequest
Contains the parameters for DescribeVpcEndpointServices.

DescribeVpcEndpointServicesResult
Contains the output of DescribeVpcEndpointServices.

DescribeVpcEndpointsRequest
Contains the parameters for DescribeVpcEndpoints.

DescribeVpcEndpointsResult
Contains the output of DescribeVpcEndpoints.

DescribeVpcPeeringConnectionsRequest
Contains the parameters for DescribeVpcPeeringConnections.

DescribeVpcPeeringConnectionsResult
Contains the output of DescribeVpcPeeringConnections.

DescribeVpcsRequest
Contains the parameters for DescribeVpcs.

DescribeVpcsResult
Contains the output of DescribeVpcs.

DescribeVpnConnectionsRequest
Contains the parameters for DescribeVpnConnections.

DescribeVpnConnectionsResult
Contains the output of DescribeVpnConnections.

DescribeVpnGatewaysRequest
Contains the parameters for DescribeVpnGateways.

DescribeVpnGatewaysResult
Contains the output of DescribeVpnGateways.

DetachClassicLinkVpcRequest
Contains the parameters for DetachClassicLinkVpc.

DetachClassicLinkVpcResult
Contains the output of DetachClassicLinkVpc.

DetachInternetGatewayRequest
Contains the parameters for DetachInternetGateway.

DetachNetworkInterfaceRequest
Contains the parameters for DetachNetworkInterface.

DetachVolumeRequest
Contains the parameters for DetachVolume.

DetachVpnGatewayRequest
Contains the parameters for DetachVpnGateway.

DhcpConfiguration
Describes a DHCP configuration option.

DhcpOptions
Describes a set of DHCP options.

DisableVgwRoutePropagationRequest
Contains the parameters for DisableVgwRoutePropagation.

DisableVpcClassicLinkDnsSupportRequest
Contains the parameters for DisableVpcClassicLinkDnsSupport.

DisableVpcClassicLinkDnsSupportResult
Contains the output of DisableVpcClassicLinkDnsSupport.

DisableVpcClassicLinkRequest
Contains the parameters for DisableVpcClassicLink.

DisableVpcClassicLinkResult
Contains the output of DisableVpcClassicLink.

DisassociateAddressRequest
Contains the parameters for DisassociateAddress.

DisassociateIamInstanceProfileRequest

DisassociateIamInstanceProfileResult

DisassociateRouteTableRequest
Contains the parameters for DisassociateRouteTable.

DisassociateSubnetCidrBlockRequest

DisassociateSubnetCidrBlockResult

DisassociateVpcCidrBlockRequest

DisassociateVpcCidrBlockResult

DiskImage
Describes a disk image.

DiskImageDescription
Describes a disk image.

DiskImageDetail
Describes a disk image.

DiskImageVolumeDescription
Describes a disk image volume.

EbsBlockDevice
Describes a block device for an EBS volume.

EbsInstanceBlockDevice
Describes a parameter used to set up an EBS volume in a block device mapping.

EbsInstanceBlockDeviceSpecification
Describes information used to set up an EBS volume specified in a block device mapping.

Ec2Client
A client for the Amazon EC2 API.

EgressOnlyInternetGateway
Describes an egress-only Internet gateway.

EnableVgwRoutePropagationRequest
Contains the parameters for EnableVgwRoutePropagation.

EnableVolumeIORequest
Contains the parameters for EnableVolumeIO.

EnableVpcClassicLinkDnsSupportRequest
Contains the parameters for EnableVpcClassicLinkDnsSupport.

EnableVpcClassicLinkDnsSupportResult
Contains the output of EnableVpcClassicLinkDnsSupport.

EnableVpcClassicLinkRequest
Contains the parameters for EnableVpcClassicLink.

EnableVpcClassicLinkResult
Contains the output of EnableVpcClassicLink.

EventInformation
Describes a Spot fleet event.

ExportTask
Describes an instance export task.

ExportToS3Task
Describes the format and location for an instance export task.

ExportToS3TaskSpecification
Describes an instance export task.

Filter
A filter name and value pair that is used to return a more specific list of results. Filters can be used to match a set of resources by various criteria, such as tags, attributes, or IDs.

FlowLog
Describes a flow log.

FpgaImage
Describes an Amazon FPGA image (AFI).

FpgaImageState
Describes the state of the bitstream generation process for an Amazon FPGA image (AFI).

GetConsoleOutputRequest
Contains the parameters for GetConsoleOutput.

GetConsoleOutputResult
Contains the output of GetConsoleOutput.

GetConsoleScreenshotRequest
Contains the parameters for the request.

GetConsoleScreenshotResult
Contains the output of the request.

GetHostReservationPurchasePreviewRequest

GetHostReservationPurchasePreviewResult

GetPasswordDataRequest
Contains the parameters for GetPasswordData.

GetPasswordDataResult
Contains the output of GetPasswordData.

GetReservedInstancesExchangeQuoteRequest
Contains the parameters for GetReservedInstanceExchangeQuote.

GetReservedInstancesExchangeQuoteResult
Contains the output of GetReservedInstancesExchangeQuote.

GroupIdentifier
Describes a security group.

HistoryRecord
Describes an event in the history of the Spot fleet request.

Host
Describes the properties of the Dedicated Host.

HostInstance
Describes an instance running on a Dedicated Host.

HostOffering
Details about the Dedicated Host Reservation offering.

HostProperties
Describes properties of a Dedicated Host.

HostReservation
Details about the Dedicated Host Reservation and associated Dedicated Hosts.

IamInstanceProfile
Describes an IAM instance profile.

IamInstanceProfileAssociation
Describes an association between an IAM instance profile and an instance.

IamInstanceProfileSpecification
Describes an IAM instance profile.

IcmpTypeCode
Describes the ICMP type and code.

IdFormat
Describes the ID format for a resource.

Image
Describes an image.

ImageAttribute
Describes an image attribute.

ImageDiskContainer
Describes the disk container object for an import image task.

ImportImageRequest
Contains the parameters for ImportImage.

ImportImageResult
Contains the output for ImportImage.

ImportImageTask
Describes an import image task.

ImportInstanceLaunchSpecification
Describes the launch specification for VM import.

ImportInstanceRequest
Contains the parameters for ImportInstance.

ImportInstanceResult
Contains the output for ImportInstance.

ImportInstanceTaskDetails
Describes an import instance task.

ImportInstanceVolumeDetailItem
Describes an import volume task.

ImportKeyPairRequest
Contains the parameters for ImportKeyPair.

ImportKeyPairResult
Contains the output of ImportKeyPair.

ImportSnapshotRequest
Contains the parameters for ImportSnapshot.

ImportSnapshotResult
Contains the output for ImportSnapshot.

ImportSnapshotTask
Describes an import snapshot task.

ImportVolumeRequest
Contains the parameters for ImportVolume.

ImportVolumeResult
Contains the output for ImportVolume.

ImportVolumeTaskDetails
Describes an import volume task.

Instance
Describes an instance.

InstanceAttribute
Describes an instance attribute.

InstanceBlockDeviceMapping
Describes a block device mapping.

InstanceBlockDeviceMappingSpecification
Describes a block device mapping entry.

InstanceCapacity
Information about the instance type that the Dedicated Host supports.

InstanceCount
Describes a Reserved Instance listing state.

InstanceExportDetails
Describes an instance to export.

InstanceIpv6Address
Describes an IPv6 address.

InstanceMonitoring
Describes the monitoring of an instance.

InstanceNetworkInterface
Describes a network interface.

InstanceNetworkInterfaceAssociation
Describes association information for an Elastic IP address (IPv4).

InstanceNetworkInterfaceAttachment
Describes a network interface attachment.

InstanceNetworkInterfaceSpecification
Describes a network interface.

InstancePrivateIpAddress
Describes a private IPv4 address.

InstanceState
Describes the current state of an instance.

InstanceStateChange
Describes an instance state change.

InstanceStatus
Describes the status of an instance.

InstanceStatusDetails
Describes the instance status.

InstanceStatusEvent
Describes a scheduled event for an instance.

InstanceStatusSummary
Describes the status of an instance.

InternetGateway
Describes an Internet gateway.

InternetGatewayAttachment
Describes the attachment of a VPC to an Internet gateway or an egress-only Internet gateway.

IpPermission
Describes a security group rule.

IpRange
Describes an IPv4 range.

Ipv6CidrBlock
Describes an IPv6 CIDR block.

Ipv6Range
[EC2-VPC only] Describes an IPv6 range.

KeyPair
Describes a key pair.

KeyPairInfo
Describes a key pair.

LaunchPermission
Describes a launch permission.

LaunchPermissionModifications
Describes a launch permission modification.

LaunchSpecification
Describes the launch specification for an instance.

ModifyHostsRequest
Contains the parameters for ModifyHosts.

ModifyHostsResult
Contains the output of ModifyHosts.

ModifyIdFormatRequest
Contains the parameters of ModifyIdFormat.

ModifyIdentityIdFormatRequest
Contains the parameters of ModifyIdentityIdFormat.

ModifyImageAttributeRequest
Contains the parameters for ModifyImageAttribute.

ModifyInstanceAttributeRequest
Contains the parameters for ModifyInstanceAttribute.

ModifyInstancePlacementRequest
Contains the parameters for ModifyInstancePlacement.

ModifyInstancePlacementResult
Contains the output of ModifyInstancePlacement.

ModifyNetworkInterfaceAttributeRequest
Contains the parameters for ModifyNetworkInterfaceAttribute.

ModifyReservedInstancesRequest
Contains the parameters for ModifyReservedInstances.

ModifyReservedInstancesResult
Contains the output of ModifyReservedInstances.

ModifySnapshotAttributeRequest
Contains the parameters for ModifySnapshotAttribute.

ModifySpotFleetRequestRequest
Contains the parameters for ModifySpotFleetRequest.

ModifySpotFleetRequestResponse
Contains the output of ModifySpotFleetRequest.

ModifySubnetAttributeRequest
Contains the parameters for ModifySubnetAttribute.

ModifyVolumeAttributeRequest
Contains the parameters for ModifyVolumeAttribute.

ModifyVolumeRequest

ModifyVolumeResult

ModifyVpcAttributeRequest
Contains the parameters for ModifyVpcAttribute.

ModifyVpcEndpointRequest
Contains the parameters for ModifyVpcEndpoint.

ModifyVpcEndpointResult
Contains the output of ModifyVpcEndpoint.

ModifyVpcPeeringConnectionOptionsRequest

ModifyVpcPeeringConnectionOptionsResult

MonitorInstancesRequest
Contains the parameters for MonitorInstances.

MonitorInstancesResult
Contains the output of MonitorInstances.

Monitoring
Describes the monitoring of an instance.

MoveAddressToVpcRequest
Contains the parameters for MoveAddressToVpc.

MoveAddressToVpcResult
Contains the output of MoveAddressToVpc.

MovingAddressStatus
Describes the status of a moving Elastic IP address.

NatGateway
Describes a NAT gateway.

NatGatewayAddress
Describes the IP addresses and network interface associated with a NAT gateway.

NetworkAcl
Describes a network ACL.

NetworkAclAssociation
Describes an association between a network ACL and a subnet.

NetworkAclEntry
Describes an entry in a network ACL.

NetworkInterface
Describes a network interface.

NetworkInterfaceAssociation
Describes association information for an Elastic IP address (IPv4 only).

NetworkInterfaceAttachment
Describes a network interface attachment.

NetworkInterfaceAttachmentChanges
Describes an attachment change.

NetworkInterfaceIpv6Address
Describes an IPv6 address associated with a network interface.

NetworkInterfacePrivateIpAddress
Describes the private IPv4 address of a network interface.

NewDhcpConfiguration

PciId
Describes the data that identifies an Amazon FPGA image (AFI) on the PCI bus.

PeeringConnectionOptions
Describes the VPC peering connection options.

PeeringConnectionOptionsRequest
The VPC peering connection options.

Placement
Describes the placement of an instance.

PlacementGroup
Describes a placement group.

PortRange
Describes a range of ports.

PrefixList
Describes prefixes for AWS services.

PrefixListId
The ID of the prefix.

PriceSchedule
Describes the price for a Reserved Instance.

PriceScheduleSpecification
Describes the price for a Reserved Instance.

PricingDetail
Describes a Reserved Instance offering.

PrivateIpAddressSpecification
Describes a secondary private IPv4 address for a network interface.

ProductCode
Describes a product code.

PropagatingVgw
Describes a virtual private gateway propagating route.

ProvisionedBandwidth
Reserved. If you need to sustain traffic greater than the documented limits, contact us through the Support Center.

Purchase
Describes the result of the purchase.

PurchaseHostReservationRequest

PurchaseHostReservationResult

PurchaseRequest
Describes a request to purchase Scheduled Instances.

PurchaseReservedInstancesOfferingRequest
Contains the parameters for PurchaseReservedInstancesOffering.

PurchaseReservedInstancesOfferingResult
Contains the output of PurchaseReservedInstancesOffering.

PurchaseScheduledInstancesRequest
Contains the parameters for PurchaseScheduledInstances.

PurchaseScheduledInstancesResult
Contains the output of PurchaseScheduledInstances.

RebootInstancesRequest
Contains the parameters for RebootInstances.

RecurringCharge
Describes a recurring charge.

Region
Describes a region.

RegisterImageRequest
Contains the parameters for RegisterImage.

RegisterImageResult
Contains the output of RegisterImage.

RejectVpcPeeringConnectionRequest
Contains the parameters for RejectVpcPeeringConnection.

RejectVpcPeeringConnectionResult
Contains the output of RejectVpcPeeringConnection.

ReleaseAddressRequest
Contains the parameters for ReleaseAddress.

ReleaseHostsRequest
Contains the parameters for ReleaseHosts.

ReleaseHostsResult
Contains the output of ReleaseHosts.

ReplaceIamInstanceProfileAssociationRequest

ReplaceIamInstanceProfileAssociationResult

ReplaceNetworkAclAssociationRequest
Contains the parameters for ReplaceNetworkAclAssociation.

ReplaceNetworkAclAssociationResult
Contains the output of ReplaceNetworkAclAssociation.

ReplaceNetworkAclEntryRequest
Contains the parameters for ReplaceNetworkAclEntry.

ReplaceRouteRequest
Contains the parameters for ReplaceRoute.

ReplaceRouteTableAssociationRequest
Contains the parameters for ReplaceRouteTableAssociation.

ReplaceRouteTableAssociationResult
Contains the output of ReplaceRouteTableAssociation.

ReportInstanceStatusRequest
Contains the parameters for ReportInstanceStatus.

RequestSpotFleetRequest
Contains the parameters for RequestSpotFleet.

RequestSpotFleetResponse
Contains the output of RequestSpotFleet.

RequestSpotInstancesRequest
Contains the parameters for RequestSpotInstances.

RequestSpotInstancesResult
Contains the output of RequestSpotInstances.

RequestSpotLaunchSpecification
Describes the launch specification for an instance.

Reservation
Describes a reservation.

ReservationValue
The cost associated with the Reserved Instance.

ReservedInstanceLimitPrice
Describes the limit price of a Reserved Instance offering.

ReservedInstanceReservationValue
The total value of the Convertible Reserved Instance.

ReservedInstances
Describes a Reserved Instance.

ReservedInstancesConfiguration
Describes the configuration settings for the modified Reserved Instances.

ReservedInstancesId
Describes the ID of a Reserved Instance.

ReservedInstancesListing
Describes a Reserved Instance listing.

ReservedInstancesModification
Describes a Reserved Instance modification.

ReservedInstancesModificationResult
Describes the modification request/s.

ReservedInstancesOffering
Describes a Reserved Instance offering.

ResetImageAttributeRequest
Contains the parameters for ResetImageAttribute.

ResetInstanceAttributeRequest
Contains the parameters for ResetInstanceAttribute.

ResetNetworkInterfaceAttributeRequest
Contains the parameters for ResetNetworkInterfaceAttribute.

ResetSnapshotAttributeRequest
Contains the parameters for ResetSnapshotAttribute.

RestoreAddressToClassicRequest
Contains the parameters for RestoreAddressToClassic.

RestoreAddressToClassicResult
Contains the output of RestoreAddressToClassic.

RevokeSecurityGroupEgressRequest
Contains the parameters for RevokeSecurityGroupEgress.

RevokeSecurityGroupIngressRequest
Contains the parameters for RevokeSecurityGroupIngress.

Route
Describes a route in a route table.

RouteTable
Describes a route table.

RouteTableAssociation
Describes an association between a route table and a subnet.

RunInstancesMonitoringEnabled
Describes the monitoring of an instance.

RunInstancesRequest
Contains the parameters for RunInstances.

RunScheduledInstancesRequest
Contains the parameters for RunScheduledInstances.

RunScheduledInstancesResult
Contains the output of RunScheduledInstances.

S3Storage
Describes the storage parameters for S3 and S3 buckets for an instance store-backed AMI.

ScheduledInstance
Describes a Scheduled Instance.

ScheduledInstanceAvailability
Describes a schedule that is available for your Scheduled Instances.

ScheduledInstanceRecurrence
Describes the recurring schedule for a Scheduled Instance.

ScheduledInstanceRecurrenceRequest
Describes the recurring schedule for a Scheduled Instance.

ScheduledInstancesBlockDeviceMapping
Describes a block device mapping for a Scheduled Instance.

ScheduledInstancesEbs
Describes an EBS volume for a Scheduled Instance.

ScheduledInstancesIamInstanceProfile
Describes an IAM instance profile for a Scheduled Instance.

ScheduledInstancesIpv6Address
Describes an IPv6 address.

ScheduledInstancesLaunchSpecification
Describes the launch specification for a Scheduled Instance.

If you are launching the Scheduled Instance in EC2-VPC, you must specify the ID of the subnet. You can specify the subnet using either SubnetId or NetworkInterface.

ScheduledInstancesMonitoring
Describes whether monitoring is enabled for a Scheduled Instance.

ScheduledInstancesNetworkInterface
Describes a network interface for a Scheduled Instance.

ScheduledInstancesPlacement
Describes the placement for a Scheduled Instance.

ScheduledInstancesPrivateIpAddressConfig
Describes a private IPv4 address for a Scheduled Instance.

SecurityGroup
Describes a security group

SecurityGroupReference
Describes a VPC with a security group that references your security group.

SlotDateTimeRangeRequest
Describes the time period for a Scheduled Instance to start its first schedule. The time period must span less than one day.

SlotStartTimeRangeRequest
Describes the time period for a Scheduled Instance to start its first schedule.

Snapshot
Describes a snapshot.

SnapshotDetail
Describes the snapshot created from the imported disk.

SnapshotDiskContainer
The disk container object for the import snapshot request.

SnapshotTaskDetail
Details about the import snapshot task.

SpotDatafeedSubscription
Describes the data feed for a Spot instance.

SpotFleetLaunchSpecification
Describes the launch specification for one or more Spot instances.

SpotFleetMonitoring
Describes whether monitoring is enabled.

SpotFleetRequestConfig
Describes a Spot fleet request.

SpotFleetRequestConfigData
Describes the configuration of a Spot fleet request.

SpotInstanceRequest
Describes a Spot instance request.

SpotInstanceStateFault
Describes a Spot instance state change.

SpotInstanceStatus
Describes the status of a Spot instance request.

SpotPlacement
Describes Spot instance placement.

SpotPrice
Describes the maximum hourly price (bid) for any Spot instance launched to fulfill the request.

StaleIpPermission
Describes a stale rule in a security group.

StaleSecurityGroup
Describes a stale security group (a security group that contains stale rules).

StartInstancesRequest
Contains the parameters for StartInstances.

StartInstancesResult
Contains the output of StartInstances.

StateReason
Describes a state change.

StopInstancesRequest
Contains the parameters for StopInstances.

StopInstancesResult
Contains the output of StopInstances.

Storage
Describes the storage location for an instance store-backed AMI.

StorageLocation
Describes a storage location in Amazon S3.

Subnet
Describes a subnet.

SubnetCidrBlockState
Describes the state of a CIDR block.

SubnetIpv6CidrBlockAssociation
Describes an IPv6 CIDR block associated with a subnet.

Tag
Describes a tag.

TagDescription
Describes a tag.

TagSpecification
The tags to apply to a resource when the resource is being created.

TargetConfiguration
Information about the Convertible Reserved Instance offering.

TargetConfigurationRequest
Details about the target configuration.

TargetReservationValue
The total value of the new Convertible Reserved Instances.

TerminateInstancesRequest
Contains the parameters for TerminateInstances.

TerminateInstancesResult
Contains the output of TerminateInstances.

UnassignIpv6AddressesRequest

UnassignIpv6AddressesResult

UnassignPrivateIpAddressesRequest
Contains the parameters for UnassignPrivateIpAddresses.

UnmonitorInstancesRequest
Contains the parameters for UnmonitorInstances.

UnmonitorInstancesResult
Contains the output of UnmonitorInstances.

UnsuccessfulItem
Information about items that were not successfully processed in a batch call.

UnsuccessfulItemError
Information about the error that occurred. For more information about errors, see Error Codes.

UserBucket
Describes the S3 bucket for the disk image.

UserBucketDetails
Describes the S3 bucket for the disk image.

UserData
Describes the user data for an instance.

UserIdGroupPair
Describes a security group and AWS account ID pair.

VgwTelemetry
Describes telemetry for a VPN tunnel.

Volume
Describes a volume.

VolumeAttachment
Describes volume attachment details.

VolumeDetail
Describes an EBS volume.

VolumeModification
Describes the modification status of an EBS volume.

If the volume has never been modified, some element values will be null.

VolumeStatusAction
Describes a volume status operation code.

VolumeStatusDetails
Describes a volume status.

VolumeStatusEvent
Describes a volume status event.

VolumeStatusInfo
Describes the status of a volume.

VolumeStatusItem
Describes the volume status.

Vpc
Describes a VPC.

VpcAttachment
Describes an attachment between a virtual private gateway and a VPC.

VpcCidrBlockState
Describes the state of a CIDR block.

VpcClassicLink
Describes whether a VPC is enabled for ClassicLink.

VpcEndpoint
Describes a VPC endpoint.

VpcIpv6CidrBlockAssociation
Describes an IPv6 CIDR block associated with a VPC.

VpcPeeringConnection
Describes a VPC peering connection.

VpcPeeringConnectionOptionsDescription
Describes the VPC peering connection options.

VpcPeeringConnectionStateReason
Describes the status of a VPC peering connection.

VpcPeeringConnectionVpcInfo
Describes a VPC in a VPC peering connection.

VpnConnection
Describes a VPN connection.

VpnConnectionOptions
Describes VPN connection options.

VpnConnectionOptionsSpecification
Describes VPN connection options.

VpnGateway
Describes a virtual private gateway.

VpnStaticRoute
Describes a static route for a VPN connection.

Enums

AcceptReservedInstancesExchangeQuoteError
Errors returned by AcceptReservedInstancesExchangeQuote

AcceptVpcPeeringConnectionError
Errors returned by AcceptVpcPeeringConnection

AllocateAddressError
Errors returned by AllocateAddress

AllocateHostsError
Errors returned by AllocateHosts

AssignIpv6AddressesError
Errors returned by AssignIpv6Addresses

AssignPrivateIpAddressesError
Errors returned by AssignPrivateIpAddresses

AssociateAddressError
Errors returned by AssociateAddress

AssociateDhcpOptionsError
Errors returned by AssociateDhcpOptions

AssociateIamInstanceProfileError
Errors returned by AssociateIamInstanceProfile

AssociateRouteTableError
Errors returned by AssociateRouteTable

AssociateSubnetCidrBlockError
Errors returned by AssociateSubnetCidrBlock

AssociateVpcCidrBlockError
Errors returned by AssociateVpcCidrBlock

AttachClassicLinkVpcError
Errors returned by AttachClassicLinkVpc

AttachInternetGatewayError
Errors returned by AttachInternetGateway

AttachNetworkInterfaceError
Errors returned by AttachNetworkInterface

AttachVolumeError
Errors returned by AttachVolume

AttachVpnGatewayError
Errors returned by AttachVpnGateway

AuthorizeSecurityGroupEgressError
Errors returned by AuthorizeSecurityGroupEgress

AuthorizeSecurityGroupIngressError
Errors returned by AuthorizeSecurityGroupIngress

BundleInstanceError
Errors returned by BundleInstance

CancelBundleTaskError
Errors returned by CancelBundleTask

CancelConversionTaskError
Errors returned by CancelConversionTask

CancelExportTaskError
Errors returned by CancelExportTask

CancelImportTaskError
Errors returned by CancelImportTask

CancelReservedInstancesListingError
Errors returned by CancelReservedInstancesListing

CancelSpotInstanceRequestsError
Errors returned by CancelSpotInstanceRequests

ConfirmProductInstanceError
Errors returned by ConfirmProductInstance

CopyImageError
Errors returned by CopyImage

CopySnapshotError
Errors returned by CopySnapshot

CreateCustomerGatewayError
Errors returned by CreateCustomerGateway

CreateDhcpOptionsError
Errors returned by CreateDhcpOptions

CreateEgressOnlyInternetGatewayError
Errors returned by CreateEgressOnlyInternetGateway

CreateFlowLogsError
Errors returned by CreateFlowLogs

CreateFpgaImageError
Errors returned by CreateFpgaImage

CreateImageError
Errors returned by CreateImage

CreateInstanceExportTaskError
Errors returned by CreateInstanceExportTask

CreateInternetGatewayError
Errors returned by CreateInternetGateway

CreateKeyPairError
Errors returned by CreateKeyPair

CreateNatGatewayError
Errors returned by CreateNatGateway

CreateNetworkAclEntryError
Errors returned by CreateNetworkAclEntry

CreateNetworkAclError
Errors returned by CreateNetworkAcl

CreateNetworkInterfaceError
Errors returned by CreateNetworkInterface

CreatePlacementGroupError
Errors returned by CreatePlacementGroup

CreateReservedInstancesListingError
Errors returned by CreateReservedInstancesListing

CreateRouteError
Errors returned by CreateRoute

CreateRouteTableError
Errors returned by CreateRouteTable

CreateSecurityGroupError
Errors returned by CreateSecurityGroup

CreateSnapshotError
Errors returned by CreateSnapshot

CreateSpotDatafeedSubscriptionError
Errors returned by CreateSpotDatafeedSubscription

CreateSubnetError
Errors returned by CreateSubnet

CreateTagsError
Errors returned by CreateTags

CreateVolumeError
Errors returned by CreateVolume

CreateVpcEndpointError
Errors returned by CreateVpcEndpoint

CreateVpcError
Errors returned by CreateVpc

CreateVpcPeeringConnectionError
Errors returned by CreateVpcPeeringConnection

CreateVpnConnectionError
Errors returned by CreateVpnConnection

CreateVpnConnectionRouteError
Errors returned by CreateVpnConnectionRoute

CreateVpnGatewayError
Errors returned by CreateVpnGateway

DeleteCustomerGatewayError
Errors returned by DeleteCustomerGateway

DeleteDhcpOptionsError
Errors returned by DeleteDhcpOptions

DeleteEgressOnlyInternetGatewayError
Errors returned by DeleteEgressOnlyInternetGateway

DeleteFlowLogsError
Errors returned by DeleteFlowLogs

DeleteInternetGatewayError
Errors returned by DeleteInternetGateway

DeleteKeyPairError
Errors returned by DeleteKeyPair

DeleteNatGatewayError
Errors returned by DeleteNatGateway

DeleteNetworkAclEntryError
Errors returned by DeleteNetworkAclEntry

DeleteNetworkAclError
Errors returned by DeleteNetworkAcl

DeleteNetworkInterfaceError
Errors returned by DeleteNetworkInterface

DeletePlacementGroupError
Errors returned by DeletePlacementGroup

DeleteRouteError
Errors returned by DeleteRoute

DeleteRouteTableError
Errors returned by DeleteRouteTable

DeleteSecurityGroupError
Errors returned by DeleteSecurityGroup

DeleteSnapshotError
Errors returned by DeleteSnapshot

DeleteSpotDatafeedSubscriptionError
Errors returned by DeleteSpotDatafeedSubscription

DeleteSubnetError
Errors returned by DeleteSubnet

DeleteTagsError
Errors returned by DeleteTags

DeleteVolumeError
Errors returned by DeleteVolume

DeleteVpcEndpointsError
Errors returned by DeleteVpcEndpoints

DeleteVpcError
Errors returned by DeleteVpc

DeleteVpcPeeringConnectionError
Errors returned by DeleteVpcPeeringConnection

DeleteVpnConnectionError
Errors returned by DeleteVpnConnection

DeleteVpnConnectionRouteError
Errors returned by DeleteVpnConnectionRoute

DeleteVpnGatewayError
Errors returned by DeleteVpnGateway

DeregisterImageError
Errors returned by DeregisterImage

DescribeAccountAttributesError
Errors returned by DescribeAccountAttributes

DescribeAddressesError
Errors returned by DescribeAddresses

DescribeAvailabilityZonesError
Errors returned by DescribeAvailabilityZones

DescribeBundleTasksError
Errors returned by DescribeBundleTasks

DescribeClassicLinkInstancesError
Errors returned by DescribeClassicLinkInstances

DescribeConversionTasksError
Errors returned by DescribeConversionTasks

DescribeCustomerGatewaysError
Errors returned by DescribeCustomerGateways

DescribeDhcpOptionsError
Errors returned by DescribeDhcpOptions

DescribeEgressOnlyInternetGatewaysError
Errors returned by DescribeEgressOnlyInternetGateways

DescribeExportTasksError
Errors returned by DescribeExportTasks

DescribeFlowLogsError
Errors returned by DescribeFlowLogs

DescribeFpgaImagesError
Errors returned by DescribeFpgaImages

DescribeHostReservationOfferingsError
Errors returned by DescribeHostReservationOfferings

DescribeHostReservationsError
Errors returned by DescribeHostReservations

DescribeHostsError
Errors returned by DescribeHosts

DescribeIamInstanceProfileAssociationsError
Errors returned by DescribeIamInstanceProfileAssociations

DescribeIdFormatError
Errors returned by DescribeIdFormat

DescribeIdentityIdFormatError
Errors returned by DescribeIdentityIdFormat

DescribeImageAttributeError
Errors returned by DescribeImageAttribute

DescribeImagesError
Errors returned by DescribeImages

DescribeImportImageTasksError
Errors returned by DescribeImportImageTasks

DescribeImportSnapshotTasksError
Errors returned by DescribeImportSnapshotTasks

DescribeInstanceAttributeError
Errors returned by DescribeInstanceAttribute

DescribeInstanceStatusError
Errors returned by DescribeInstanceStatus

DescribeInstancesError
Errors returned by DescribeInstances

DescribeInternetGatewaysError
Errors returned by DescribeInternetGateways

DescribeKeyPairsError
Errors returned by DescribeKeyPairs

DescribeMovingAddressesError
Errors returned by DescribeMovingAddresses

DescribeNatGatewaysError
Errors returned by DescribeNatGateways

DescribeNetworkAclsError
Errors returned by DescribeNetworkAcls

DescribeNetworkInterfaceAttributeError
Errors returned by DescribeNetworkInterfaceAttribute

DescribeNetworkInterfacesError
Errors returned by DescribeNetworkInterfaces

DescribePlacementGroupsError
Errors returned by DescribePlacementGroups

DescribePrefixListsError
Errors returned by DescribePrefixLists

DescribeRegionsError
Errors returned by DescribeRegions

DescribeReservedInstancesError
Errors returned by DescribeReservedInstances

DescribeReservedInstancesListingsError
Errors returned by DescribeReservedInstancesListings

DescribeReservedInstancesModificationsError
Errors returned by DescribeReservedInstancesModifications

DescribeReservedInstancesOfferingsError
Errors returned by DescribeReservedInstancesOfferings

DescribeRouteTablesError
Errors returned by DescribeRouteTables

DescribeScheduledInstanceAvailabilityError
Errors returned by DescribeScheduledInstanceAvailability

DescribeScheduledInstancesError
Errors returned by DescribeScheduledInstances

DescribeSecurityGroupReferencesError
Errors returned by DescribeSecurityGroupReferences

DescribeSecurityGroupsError
Errors returned by DescribeSecurityGroups

DescribeSnapshotAttributeError
Errors returned by DescribeSnapshotAttribute

DescribeSnapshotsError
Errors returned by DescribeSnapshots

DescribeSpotDatafeedSubscriptionError
Errors returned by DescribeSpotDatafeedSubscription

DescribeSpotFleetInstancesError
Errors returned by DescribeSpotFleetInstances

DescribeSpotFleetRequestHistoryError
Errors returned by DescribeSpotFleetRequestHistory

DescribeSpotFleetRequestsError
Errors returned by DescribeSpotFleetRequests

DescribeSpotInstanceRequestsError
Errors returned by DescribeSpotInstanceRequests

DescribeSpotPriceHistoryError
Errors returned by DescribeSpotPriceHistory

DescribeStaleSecurityGroupsError
Errors returned by DescribeStaleSecurityGroups

DescribeSubnetsError
Errors returned by DescribeSubnets

DescribeTagsError
Errors returned by DescribeTags

DescribeVolumeAttributeError
Errors returned by DescribeVolumeAttribute

DescribeVolumeStatusError
Errors returned by DescribeVolumeStatus

DescribeVolumesError
Errors returned by DescribeVolumes

DescribeVolumesModificationsError
Errors returned by DescribeVolumesModifications

DescribeVpcAttributeError
Errors returned by DescribeVpcAttribute

DescribeVpcClassicLinkDnsSupportError
Errors returned by DescribeVpcClassicLinkDnsSupport

DescribeVpcClassicLinkError
Errors returned by DescribeVpcClassicLink

DescribeVpcEndpointServicesError
Errors returned by DescribeVpcEndpointServices

DescribeVpcEndpointsError
Errors returned by DescribeVpcEndpoints

DescribeVpcPeeringConnectionsError
Errors returned by DescribeVpcPeeringConnections

DescribeVpcsError
Errors returned by DescribeVpcs

DescribeVpnConnectionsError
Errors returned by DescribeVpnConnections

DescribeVpnGatewaysError
Errors returned by DescribeVpnGateways

DetachClassicLinkVpcError
Errors returned by DetachClassicLinkVpc

DetachInternetGatewayError
Errors returned by DetachInternetGateway

DetachNetworkInterfaceError
Errors returned by DetachNetworkInterface

DetachVolumeError
Errors returned by DetachVolume

DetachVpnGatewayError
Errors returned by DetachVpnGateway

DisableVgwRoutePropagationError
Errors returned by DisableVgwRoutePropagation

DisableVpcClassicLinkDnsSupportError
Errors returned by DisableVpcClassicLinkDnsSupport

DisableVpcClassicLinkError
Errors returned by DisableVpcClassicLink

DisassociateAddressError
Errors returned by DisassociateAddress

DisassociateIamInstanceProfileError
Errors returned by DisassociateIamInstanceProfile

DisassociateRouteTableError
Errors returned by DisassociateRouteTable

DisassociateSubnetCidrBlockError
Errors returned by DisassociateSubnetCidrBlock

DisassociateVpcCidrBlockError
Errors returned by DisassociateVpcCidrBlock

EC2CancelSpotFleetRequestsError
Errors returned by CancelSpotFleetRequests

EnableVgwRoutePropagationError
Errors returned by EnableVgwRoutePropagation

EnableVolumeIOError
Errors returned by EnableVolumeIO

EnableVpcClassicLinkDnsSupportError
Errors returned by EnableVpcClassicLinkDnsSupport

EnableVpcClassicLinkError
Errors returned by EnableVpcClassicLink

GetConsoleOutputError
Errors returned by GetConsoleOutput

GetConsoleScreenshotError
Errors returned by GetConsoleScreenshot

GetHostReservationPurchasePreviewError
Errors returned by GetHostReservationPurchasePreview

GetPasswordDataError
Errors returned by GetPasswordData

GetReservedInstancesExchangeQuoteError
Errors returned by GetReservedInstancesExchangeQuote

ImportImageError
Errors returned by ImportImage

ImportInstanceError
Errors returned by ImportInstance

ImportKeyPairError
Errors returned by ImportKeyPair

ImportSnapshotError
Errors returned by ImportSnapshot

ImportVolumeError
Errors returned by ImportVolume

ModifyHostsError
Errors returned by ModifyHosts

ModifyIdFormatError
Errors returned by ModifyIdFormat

ModifyIdentityIdFormatError
Errors returned by ModifyIdentityIdFormat

ModifyImageAttributeError
Errors returned by ModifyImageAttribute

ModifyInstanceAttributeError
Errors returned by ModifyInstanceAttribute

ModifyInstancePlacementError
Errors returned by ModifyInstancePlacement

ModifyNetworkInterfaceAttributeError
Errors returned by ModifyNetworkInterfaceAttribute

ModifyReservedInstancesError
Errors returned by ModifyReservedInstances

ModifySnapshotAttributeError
Errors returned by ModifySnapshotAttribute

ModifySpotFleetRequestError
Errors returned by ModifySpotFleetRequest

ModifySubnetAttributeError
Errors returned by ModifySubnetAttribute

ModifyVolumeAttributeError
Errors returned by ModifyVolumeAttribute

ModifyVolumeError
Errors returned by ModifyVolume

ModifyVpcAttributeError
Errors returned by ModifyVpcAttribute

ModifyVpcEndpointError
Errors returned by ModifyVpcEndpoint

ModifyVpcPeeringConnectionOptionsError
Errors returned by ModifyVpcPeeringConnectionOptions

MonitorInstancesError
Errors returned by MonitorInstances

MoveAddressToVpcError
Errors returned by MoveAddressToVpc

PurchaseHostReservationError
Errors returned by PurchaseHostReservation

PurchaseReservedInstancesOfferingError
Errors returned by PurchaseReservedInstancesOffering

PurchaseScheduledInstancesError
Errors returned by PurchaseScheduledInstances

RebootInstancesError
Errors returned by RebootInstances

RegisterImageError
Errors returned by RegisterImage

RejectVpcPeeringConnectionError
Errors returned by RejectVpcPeeringConnection

ReleaseAddressError
Errors returned by ReleaseAddress

ReleaseHostsError
Errors returned by ReleaseHosts

ReplaceIamInstanceProfileAssociationError
Errors returned by ReplaceIamInstanceProfileAssociation

ReplaceNetworkAclAssociationError
Errors returned by ReplaceNetworkAclAssociation

ReplaceNetworkAclEntryError
Errors returned by ReplaceNetworkAclEntry

ReplaceRouteError
Errors returned by ReplaceRoute

ReplaceRouteTableAssociationError
Errors returned by ReplaceRouteTableAssociation

ReportInstanceStatusError
Errors returned by ReportInstanceStatus

RequestSpotFleetError
Errors returned by RequestSpotFleet

RequestSpotInstancesError
Errors returned by RequestSpotInstances

ResetImageAttributeError
Errors returned by ResetImageAttribute

ResetInstanceAttributeError
Errors returned by ResetInstanceAttribute

ResetNetworkInterfaceAttributeError
Errors returned by ResetNetworkInterfaceAttribute

ResetSnapshotAttributeError
Errors returned by ResetSnapshotAttribute

RestoreAddressToClassicError
Errors returned by RestoreAddressToClassic

RevokeSecurityGroupEgressError
Errors returned by RevokeSecurityGroupEgress

RevokeSecurityGroupIngressError
Errors returned by RevokeSecurityGroupIngress

RunInstancesError
Errors returned by RunInstances

RunScheduledInstancesError
Errors returned by RunScheduledInstances

StartInstancesError
Errors returned by StartInstances

StopInstancesError
Errors returned by StopInstances

TerminateInstancesError
Errors returned by TerminateInstances

UnassignIpv6AddressesError
Errors returned by UnassignIpv6Addresses

UnassignPrivateIpAddressesError
Errors returned by UnassignPrivateIpAddresses

UnmonitorInstancesError
Errors returned by UnmonitorInstances

Traits

Ec2
Trait representing the capabilities of the Amazon EC2 API. Amazon EC2 clients implement this trait.