Struct rusoto_ec2::CopyImageRequest

pub struct CopyImageRequest {
    pub client_token: Option<String>,
    pub description: Option<String>,
    pub dry_run: Option<bool>,
    pub encrypted: Option<bool>,
    pub kms_key_id: Option<String>,
    pub name: String,
    pub source_image_id: String,
    pub source_region: String,
}

Contains the parameters for CopyImage.

Fields
client_token: Option<String>

Unique, case-sensitive identifier you provide to ensure idempotency of the request. For more information, see How to Ensure Idempotency in the Amazon Elastic Compute Cloud User Guide.

description: Option<String>

A description for the new AMI in the destination region.

dry_run: Option<bool>

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

encrypted: Option<bool>

Specifies whether the destination snapshots of the copied image should be encrypted. The default CMK for EBS is used unless a non-default AWS Key Management Service (AWS KMS) CMK is specified with KmsKeyId. For more information, see Amazon EBS Encryption in the Amazon Elastic Compute Cloud User Guide.

kms_key_id: Option<String>

The full ARN of the AWS Key Management Service (AWS KMS) CMK to use when encrypting the snapshots of an image during a copy operation. This parameter is only required if you want to use a non-default CMK; if this parameter is not specified, the default CMK for EBS is used. The ARN contains the arn:aws:kms namespace, followed by the region of the CMK, the AWS account ID of the CMK owner, the key namespace, and then the CMK ID. For example, arn:aws:kms:us-east-1:012345678910:key/abcd1234-a123-456a-a12b-a123b4cd56ef. The specified CMK must exist in the region that the snapshot is being copied to. If a KmsKeyId is specified, the Encrypted flag must also be set.

name: String

The name of the new AMI in the destination region.

source_image_id: String

The ID of the AMI to copy.

source_region: String

The name of the region that contains the AMI to copy.