Struct rusoto_ec2::CreateNetworkInterfaceRequest

pub struct CreateNetworkInterfaceRequest {
    pub description: Option<String>,
    pub dry_run: Option<bool>,
    pub groups: Option<Vec<String>>,
    pub ipv_6_address_count: Option<i64>,
    pub ipv_6_addresses: Option<Vec<InstanceIpv6Address>>,
    pub private_ip_address: Option<String>,
    pub private_ip_addresses: Option<Vec<PrivateIpAddressSpecification>>,
    pub secondary_private_ip_address_count: Option<i64>,
    pub subnet_id: String,
}

Contains the parameters for CreateNetworkInterface.

Fields
description: Option<String>

A description for the network interface.

dry_run: Option<bool>

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

groups: Option<Vec<String>>

The IDs of one or more security groups.

ipv_6_address_count: Option<i64>

The number of IPv6 addresses to assign to a network interface. Amazon EC2 automatically selects the IPv6 addresses from the subnet range. You can't use this option if specifying specific IPv6 addresses. If your subnet has the AssignIpv6AddressOnCreation attribute set to true, you can specify 0 to override this setting.

ipv_6_addresses: Option<Vec<InstanceIpv6Address>>

One or more specific IPv6 addresses from the IPv6 CIDR block range of your subnet. You can't use this option if you're specifying a number of IPv6 addresses.

private_ip_address: Option<String>

The primary private IPv4 address of the network interface. If you don't specify an IPv4 address, Amazon EC2 selects one for you from the subnet's IPv4 CIDR range. If you specify an IP address, you cannot indicate any IP addresses specified in privateIpAddresses as primary (only one IP address can be designated as primary).

private_ip_addresses: Option<Vec<PrivateIpAddressSpecification>>

One or more private IPv4 addresses.

secondary_private_ip_address_count: Option<i64>

The number of secondary private IPv4 addresses to assign to a network interface. When you specify a number of secondary IPv4 addresses, Amazon EC2 selects these IP addresses within the subnet's IPv4 CIDR range. You can't specify this option and specify more than one private IP address using privateIpAddresses.

The number of IP addresses you can assign to a network interface varies by instance type. For more information, see IP Addresses Per ENI Per Instance Type in the Amazon Virtual Private Cloud User Guide.

subnet_id: String

The ID of the subnet to associate with the network interface.