Struct rusoto_ec2::CreateVolumeRequest

pub struct CreateVolumeRequest {
    pub availability_zone: String,
    pub dry_run: Option<bool>,
    pub encrypted: Option<bool>,
    pub iops: Option<i64>,
    pub kms_key_id: Option<String>,
    pub size: Option<i64>,
    pub snapshot_id: Option<String>,
    pub tag_specifications: Option<Vec<TagSpecification>>,
    pub volume_type: Option<String>,
}

Contains the parameters for CreateVolume.

Fields
availability_zone: String

The Availability Zone in which to create the volume. Use DescribeAvailabilityZones to list the Availability Zones that are currently available to you.

dry_run: Option<bool>

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

encrypted: Option<bool>

Specifies whether the volume should be encrypted. Encrypted Amazon EBS volumes may only be attached to instances that support Amazon EBS encryption. Volumes that are created from encrypted snapshots are automatically encrypted. There is no way to create an encrypted volume from an unencrypted snapshot or vice versa. If your AMI uses encrypted volumes, you can only launch it on supported instance types. For more information, see Amazon EBS Encryption in the Amazon Elastic Compute Cloud User Guide.

iops: Option<i64>

Only valid for Provisioned IOPS SSD volumes. The number of I/O operations per second (IOPS) to provision for the volume, with a maximum ratio of 50 IOPS/GiB.

Constraint: Range is 100 to 20000 for Provisioned IOPS SSD volumes

kms_key_id: Option<String>

The full ARN of the AWS Key Management Service (AWS KMS) customer master key (CMK) to use when creating the encrypted volume. This parameter is only required if you want to use a non-default CMK; if this parameter is not specified, the default CMK for EBS is used. The ARN contains the arn:aws:kms namespace, followed by the region of the CMK, the AWS account ID of the CMK owner, the key namespace, and then the CMK ID. For example, arn:aws:kms:us-east-1:012345678910:key/abcd1234-a123-456a-a12b-a123b4cd56ef. If a KmsKeyId is specified, the Encrypted flag must also be set.

size: Option<i64>

The size of the volume, in GiBs.

Constraints: 1-16384 for gp2, 4-16384 for io1, 500-16384 for st1, 500-16384 for sc1, and 1-1024 for standard. If you specify a snapshot, the volume size must be equal to or larger than the snapshot size.

Default: If you're creating the volume from a snapshot and don't specify a volume size, the default is the snapshot size.

snapshot_id: Option<String>

The snapshot from which to create the volume.

tag_specifications: Option<Vec<TagSpecification>>

The tags to apply to the volume during creation.

volume_type: Option<String>

The volume type. This can be gp2 for General Purpose SSD, io1 for Provisioned IOPS SSD, st1 for Throughput Optimized HDD, sc1 for Cold HDD, or standard for Magnetic volumes.

Default: standard