Struct rusoto_ec2::CreateVpcEndpointRequest

pub struct CreateVpcEndpointRequest {
    pub client_token: Option<String>,
    pub dry_run: Option<bool>,
    pub policy_document: Option<String>,
    pub private_dns_enabled: Option<bool>,
    pub route_table_ids: Option<Vec<String>>,
    pub security_group_ids: Option<Vec<String>>,
    pub service_name: String,
    pub subnet_ids: Option<Vec<String>>,
    pub vpc_endpoint_type: Option<String>,
    pub vpc_id: String,
}

Contains the parameters for CreateVpcEndpoint.

Fields
client_token: Option<String>

Unique, case-sensitive identifier you provide to ensure the idempotency of the request. For more information, see How to Ensure Idempotency.

dry_run: Option<bool>

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

policy_document: Option<String>

(Gateway endpoint) A policy to attach to the endpoint that controls access to the service. The policy must be in valid JSON format. If this parameter is not specified, we attach a default policy that allows full access to the service.

private_dns_enabled: Option<bool>

(Interface endpoint) Indicate whether to associate a private hosted zone with the specified VPC. The private hosted zone contains a record set for the default public DNS name for the service for the region (for example, kinesis.us-east-1.amazonaws.com) which resolves to the private IP addresses of the endpoint network interfaces in the VPC. This enables you to make requests to the default public DNS name for the service instead of the public DNS names that are automatically generated by the VPC endpoint service.

To use a private hosted zone, you must set the following VPC attributes to true: enableDnsHostnames and enableDnsSupport. Use ModifyVpcAttribute to set the VPC attributes.

Default: true

route_table_ids: Option<Vec<String>>

(Gateway endpoint) One or more route table IDs.

security_group_ids: Option<Vec<String>>

(Interface endpoint) The ID of one or more security groups to associate with the endpoint network interface.

service_name: String

The service name. To get a list of available services, use the DescribeVpcEndpointServices request.

subnet_ids: Option<Vec<String>>

(Interface endpoint) The ID of one or more subnets in which to create an endpoint network interface.

vpc_endpoint_type: Option<String>

The type of endpoint.

Default: Gateway

vpc_id: String

The ID of the VPC in which the endpoint will be used.