Struct rusoto_ec2::DescribeLaunchTemplateVersionsRequest

pub struct DescribeLaunchTemplateVersionsRequest {
    pub dry_run: Option<bool>,
    pub filters: Option<Vec<Filter>>,
    pub launch_template_id: Option<String>,
    pub launch_template_name: Option<String>,
    pub max_results: Option<i64>,
    pub max_version: Option<String>,
    pub min_version: Option<String>,
    pub next_token: Option<String>,
    pub versions: Option<Vec<String>>,
}

Fields
dry_run: Option<bool>

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

filters: Option<Vec<Filter>>

One or more filters.

create-time - The time the launch template version was created.

ebs-optimized - A boolean that indicates whether the instance is optimized for Amazon EBS I/O.

iam-instance-profile - The ARN of the IAM instance profile.

image-id - The ID of the AMI.

instance-type - The instance type.

is-default-version - A boolean that indicates whether the launch template version is the default version.

kernel-id - The kernel ID.

ram-disk-id - The RAM disk ID.

launch_template_id: Option<String>

The ID of the launch template. You must specify either the launch template ID or launch template name in the request.

launch_template_name: Option<String>

The name of the launch template. You must specify either the launch template ID or launch template name in the request.

max_results: Option<i64>

The maximum number of results to return in a single call. To retrieve the remaining results, make another call with the returned NextToken value. This value can be between 5 and 1000.

max_version: Option<String>

The version number up to which to describe launch template versions.

min_version: Option<String>

The version number after which to describe launch template versions.

next_token: Option<String>

The token to request the next page of results.

versions: Option<Vec<String>>

One or more versions of the launch template.