Struct rusoto_ec2::DescribeNetworkInterfacesRequest

pub struct DescribeNetworkInterfacesRequest {
    pub dry_run: Option<bool>,
    pub filters: Option<Vec<Filter>>,
    pub network_interface_ids: Option<Vec<String>>,
}

Contains the parameters for DescribeNetworkInterfaces.

Fields
dry_run: Option<bool>

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

filters: Option<Vec<Filter>>

One or more filters.

addresses.private-ip-address - The private IPv4 addresses associated with the network interface.

addresses.primary - Whether the private IPv4 address is the primary IP address associated with the network interface.

addresses.association.public-ip - The association ID returned when the network interface was associated with the Elastic IP address (IPv4).

addresses.association.owner-id - The owner ID of the addresses associated with the network interface.

association.association-id - The association ID returned when the network interface was associated with an IPv4 address.

association.allocation-id - The allocation ID returned when you allocated the Elastic IP address (IPv4) for your network interface.

association.ip-owner-id - The owner of the Elastic IP address (IPv4) associated with the network interface.

association.public-ip - The address of the Elastic IP address (IPv4) bound to the network interface.

association.public-dns-name - The public DNS name for the network interface (IPv4).

attachment.attachment-id - The ID of the interface attachment.

attachment.attach.time - The time that the network interface was attached to an instance.

attachment.delete-on-termination - Indicates whether the attachment is deleted when an instance is terminated.

attachment.device-index - The device index to which the network interface is attached.

attachment.instance-id - The ID of the instance to which the network interface is attached.

attachment.instance-owner-id - The owner ID of the instance to which the network interface is attached.

attachment.nat-gateway-id - The ID of the NAT gateway to which the network interface is attached.

attachment.status - The status of the attachment (attaching | attached | detaching | detached).

availability-zone - The Availability Zone of the network interface.

description - The description of the network interface.

group-id - The ID of a security group associated with the network interface.

group-name - The name of a security group associated with the network interface.

ipv6-addresses.ipv6-address - An IPv6 address associated with the network interface.

mac-address - The MAC address of the network interface.

network-interface-id - The ID of the network interface.

owner-id - The AWS account ID of the network interface owner.

private-ip-address - The private IPv4 address or addresses of the network interface.

private-dns-name - The private DNS name of the network interface (IPv4).

requester-id - The ID of the entity that launched the instance on your behalf (for example, AWS Management Console, Auto Scaling, and so on).

requester-managed - Indicates whether the network interface is being managed by an AWS service (for example, AWS Management Console, Auto Scaling, and so on).

source-desk-check - Indicates whether the network interface performs source/destination checking. A value of true means checking is enabled, and false means checking is disabled. The value must be false for the network interface to perform network address translation (NAT) in your VPC.

status - The status of the network interface. If the network interface is not attached to an instance, the status is available; if a network interface is attached to an instance the status is in-use.

subnet-id - The ID of the subnet for the network interface.

tag:key=value - The key/value combination of a tag assigned to the resource. Specify the key of the tag in the filter name and the value of the tag in the filter value. For example, for the tag Purpose=X, specify tag:Purpose for the filter name and X for the filter value.

tag-key - The key of a tag assigned to the resource. This filter is independent of the tag-value filter. For example, if you use both the filter "tag-key=Purpose" and the filter "tag-value=X", you get any resources assigned both the tag key Purpose (regardless of what the tag's value is), and the tag value X (regardless of what the tag's key is). If you want to list only resources where Purpose is X, see the tag:key=value filter.

tag-value - The value of a tag assigned to the resource. This filter is independent of the tag-key filter.

vpc-id - The ID of the VPC for the network interface.

network_interface_ids: Option<Vec<String>>

One or more network interface IDs.

Default: Describes all your network interfaces.