Struct rusoto_ec2::DescribeScheduledInstanceAvailabilityRequest

pub struct DescribeScheduledInstanceAvailabilityRequest {
    pub dry_run: Option<bool>,
    pub filters: Option<Vec<Filter>>,
    pub first_slot_start_time_range: SlotDateTimeRangeRequest,
    pub max_results: Option<i64>,
    pub max_slot_duration_in_hours: Option<i64>,
    pub min_slot_duration_in_hours: Option<i64>,
    pub next_token: Option<String>,
    pub recurrence: ScheduledInstanceRecurrenceRequest,
}

Contains the parameters for DescribeScheduledInstanceAvailability.

Fields
dry_run: Option<bool>

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

filters: Option<Vec<Filter>>

One or more filters.

availability-zone - The Availability Zone (for example, us-west-2a).

instance-type - The instance type (for example, c4.large).

network-platform - The network platform (EC2-Classic or EC2-VPC).

platform - The platform (Linux/UNIX or Windows).

first_slot_start_time_range: SlotDateTimeRangeRequest

The time period for the first schedule to start.

max_results: Option<i64>

The maximum number of results to return in a single call. This value can be between 5 and 300. The default value is 300. To retrieve the remaining results, make another call with the returned NextToken value.

max_slot_duration_in_hours: Option<i64>

The maximum available duration, in hours. This value must be greater than MinSlotDurationInHours and less than 1,720.

min_slot_duration_in_hours: Option<i64>

The minimum available duration, in hours. The minimum required duration is 1,200 hours per year. For example, the minimum daily schedule is 4 hours, the minimum weekly schedule is 24 hours, and the minimum monthly schedule is 100 hours.

next_token: Option<String>

The token for the next set of results.

recurrence: ScheduledInstanceRecurrenceRequest

The schedule recurrence.