Struct rusoto_ec2::DescribeSubnetsRequest

pub struct DescribeSubnetsRequest {
    pub dry_run: Option<bool>,
    pub filters: Option<Vec<Filter>>,
    pub subnet_ids: Option<Vec<String>>,
}

Contains the parameters for DescribeSubnets.

Fields
dry_run: Option<bool>

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

filters: Option<Vec<Filter>>

One or more filters.

availabilityZone - The Availability Zone for the subnet. You can also use availability-zone as the filter name.

available-ip-address-count - The number of IPv4 addresses in the subnet that are available.

cidrBlock - The IPv4 CIDR block of the subnet. The CIDR block you specify must exactly match the subnet's CIDR block for information to be returned for the subnet. You can also use cidr or cidr-block as the filter names.

defaultForAz - Indicates whether this is the default subnet for the Availability Zone. You can also use default-for-az as the filter name.

ipv6-cidr-block-association.ipv6-cidr-block - An IPv6 CIDR block associated with the subnet.

ipv6-cidr-block-association.association-id - An association ID for an IPv6 CIDR block associated with the subnet.

ipv6-cidr-block-association.state - The state of an IPv6 CIDR block associated with the subnet.

state - The state of the subnet (pending | available).

subnet-id - The ID of the subnet.

tag:key=value - The key/value combination of a tag assigned to the resource. Specify the key of the tag in the filter name and the value of the tag in the filter value. For example, for the tag Purpose=X, specify tag:Purpose for the filter name and X for the filter value.

tag-key - The key of a tag assigned to the resource. This filter is independent of the tag-value filter. For example, if you use both the filter "tag-key=Purpose" and the filter "tag-value=X", you get any resources assigned both the tag key Purpose (regardless of what the tag's value is), and the tag value X (regardless of what the tag's key is). If you want to list only resources where Purpose is X, see the tag:key=value filter.

tag-value - The value of a tag assigned to the resource. This filter is independent of the tag-key filter.

vpc-id - The ID of the VPC for the subnet.

subnet_ids: Option<Vec<String>>

One or more subnet IDs.

Default: Describes all your subnets.