Struct rusoto_ec2::DescribeVolumesRequest

pub struct DescribeVolumesRequest {
    pub dry_run: Option<bool>,
    pub filters: Option<Vec<Filter>>,
    pub max_results: Option<i64>,
    pub next_token: Option<String>,
    pub volume_ids: Option<Vec<String>>,
}

Contains the parameters for DescribeVolumes.

Fields
dry_run: Option<bool>

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

filters: Option<Vec<Filter>>

One or more filters.

attachment.attach-time - The time stamp when the attachment initiated.

attachment.delete-on-termination - Whether the volume is deleted on instance termination.

attachment.device - The device name that is exposed to the instance (for example, /dev/sda1).

attachment.instance-id - The ID of the instance the volume is attached to.

attachment.status - The attachment state (attaching | attached | detaching | detached).

availability-zone - The Availability Zone in which the volume was created.

create-time - The time stamp when the volume was created.

encrypted - The encryption status of the volume.

size - The size of the volume, in GiB.

snapshot-id - The snapshot from which the volume was created.

status - The status of the volume (creating | available | in-use | deleting | deleted | error).

tag:key=value - The key/value combination of a tag assigned to the resource. Specify the key of the tag in the filter name and the value of the tag in the filter value. For example, for the tag Purpose=X, specify tag:Purpose for the filter name and X for the filter value.

tag-key - The key of a tag assigned to the resource. This filter is independent of the tag-value filter. For example, if you use both the filter "tag-key=Purpose" and the filter "tag-value=X", you get any resources assigned both the tag key Purpose (regardless of what the tag's value is), and the tag value X (regardless of what the tag's key is). If you want to list only resources where Purpose is X, see the tag:key=value filter.

tag-value - The value of a tag assigned to the resource. This filter is independent of the tag-key filter.

volume-id - The volume ID.

volume-type - The Amazon EBS volume type. This can be gp2 for General Purpose SSD, io1 for Provisioned IOPS SSD, st1 for Throughput Optimized HDD, sc1 for Cold HDD, or standard for Magnetic volumes.

max_results: Option<i64>

The maximum number of volume results returned by DescribeVolumes in paginated output. When this parameter is used, DescribeVolumes only returns MaxResults results in a single page along with a NextToken response element. The remaining results of the initial request can be seen by sending another DescribeVolumes request with the returned NextToken value. This value can be between 5 and 500; if MaxResults is given a value larger than 500, only 500 results are returned. If this parameter is not used, then DescribeVolumes returns all results. You cannot specify this parameter and the volume IDs parameter in the same request.

next_token: Option<String>

The NextToken value returned from a previous paginated DescribeVolumes request where MaxResults was used and the results exceeded the value of that parameter. Pagination continues from the end of the previous results that returned the NextToken value. This value is null when there are no more results to return.

volume_ids: Option<Vec<String>>

One or more volume IDs.