Struct rusoto_ec2::Image

pub struct Image {
    pub architecture: Option<String>,
    pub block_device_mappings: Option<Vec<BlockDeviceMapping>>,
    pub creation_date: Option<String>,
    pub description: Option<String>,
    pub ena_support: Option<bool>,
    pub hypervisor: Option<String>,
    pub image_id: Option<String>,
    pub image_location: Option<String>,
    pub image_owner_alias: Option<String>,
    pub image_type: Option<String>,
    pub kernel_id: Option<String>,
    pub name: Option<String>,
    pub owner_id: Option<String>,
    pub platform: Option<String>,
    pub product_codes: Option<Vec<ProductCode>>,
    pub public: Option<bool>,
    pub ramdisk_id: Option<String>,
    pub root_device_name: Option<String>,
    pub root_device_type: Option<String>,
    pub sriov_net_support: Option<String>,
    pub state: Option<String>,
    pub state_reason: Option<StateReason>,
    pub tags: Option<Vec<Tag>>,
    pub virtualization_type: Option<String>,
}

Describes an image.

Fields
architecture: Option<String>

The architecture of the image.

block_device_mappings: Option<Vec<BlockDeviceMapping>>

Any block device mapping entries.

creation_date: Option<String>

The date and time the image was created.

description: Option<String>

The description of the AMI that was provided during image creation.

ena_support: Option<bool>

Specifies whether enhanced networking with ENA is enabled.

hypervisor: Option<String>

The hypervisor type of the image.

image_id: Option<String>

The ID of the AMI.

image_location: Option<String>

The location of the AMI.

image_owner_alias: Option<String>

The AWS account alias (for example, amazon, self) or the AWS account ID of the AMI owner.

image_type: Option<String>

The type of image.

kernel_id: Option<String>

The kernel associated with the image, if any. Only applicable for machine images.

name: Option<String>

The name of the AMI that was provided during image creation.

owner_id: Option<String>

The AWS account ID of the image owner.

platform: Option<String>

The value is Windows for Windows AMIs; otherwise blank.

product_codes: Option<Vec<ProductCode>>

Any product codes associated with the AMI.

public: Option<bool>

Indicates whether the image has public launch permissions. The value is true if this image has public launch permissions or false if it has only implicit and explicit launch permissions.

ramdisk_id: Option<String>

The RAM disk associated with the image, if any. Only applicable for machine images.

root_device_name: Option<String>

The device name of the root device (for example, /dev/sda1 or /dev/xvda).

root_device_type: Option<String>

The type of root device used by the AMI. The AMI can use an EBS volume or an instance store volume.

sriov_net_support: Option<String>

Specifies whether enhanced networking with the Intel 82599 Virtual Function interface is enabled.

state: Option<String>

The current state of the AMI. If the state is available, the image is successfully registered and can be used to launch an instance.

state_reason: Option<StateReason>

The reason for the state change.

tags: Option<Vec<Tag>>

Any tags assigned to the image.

virtualization_type: Option<String>

The type of virtualization of the AMI.