Struct rusoto_ec2::ImportImageRequest

pub struct ImportImageRequest {
    pub architecture: Option<String>,
    pub client_data: Option<ClientData>,
    pub client_token: Option<String>,
    pub description: Option<String>,
    pub disk_containers: Option<Vec<ImageDiskContainer>>,
    pub dry_run: Option<bool>,
    pub hypervisor: Option<String>,
    pub license_type: Option<String>,
    pub platform: Option<String>,
    pub role_name: Option<String>,
}

Contains the parameters for ImportImage.

Fields
architecture: Option<String>

The architecture of the virtual machine.

Valid values: i386 | x86_64

client_data: Option<ClientData>

The client-specific data.

client_token: Option<String>

The token to enable idempotency for VM import requests.

description: Option<String>

A description string for the import image task.

disk_containers: Option<Vec<ImageDiskContainer>>

Information about the disk containers.

dry_run: Option<bool>

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

hypervisor: Option<String>

The target hypervisor platform.

Valid values: xen

license_type: Option<String>

The license type to be used for the Amazon Machine Image (AMI) after importing.

Note: You may only use BYOL if you have existing licenses with rights to use these licenses in a third party cloud like AWS. For more information, see Prerequisites in the VM Import/Export User Guide.

Valid values: AWS | BYOL

platform: Option<String>

The operating system of the virtual machine.

Valid values: Windows | Linux

role_name: Option<String>

The name of the role to use when not using the default role, 'vmimport'.