Struct rusoto_ec2::ImportImageTask

pub struct ImportImageTask {
    pub architecture: Option<String>,
    pub description: Option<String>,
    pub hypervisor: Option<String>,
    pub image_id: Option<String>,
    pub import_task_id: Option<String>,
    pub license_type: Option<String>,
    pub platform: Option<String>,
    pub progress: Option<String>,
    pub snapshot_details: Option<Vec<SnapshotDetail>>,
    pub status: Option<String>,
    pub status_message: Option<String>,
}

Describes an import image task.

Fields
architecture: Option<String>

The architecture of the virtual machine.

Valid values: i386 | x86_64

description: Option<String>

A description of the import task.

hypervisor: Option<String>

The target hypervisor for the import task.

Valid values: xen

image_id: Option<String>

The ID of the Amazon Machine Image (AMI) of the imported virtual machine.

import_task_id: Option<String>

The ID of the import image task.

license_type: Option<String>

The license type of the virtual machine.

platform: Option<String>

The description string for the import image task.

progress: Option<String>

The percentage of progress of the import image task.

snapshot_details: Option<Vec<SnapshotDetail>>

Information about the snapshots.

status: Option<String>

A brief status for the import image task.

status_message: Option<String>

A descriptive status message for the import image task.