Struct rusoto_ec2::Instance

pub struct Instance {
    pub ami_launch_index: Option<i64>,
    pub architecture: Option<String>,
    pub block_device_mappings: Option<Vec<InstanceBlockDeviceMapping>>,
    pub client_token: Option<String>,
    pub ebs_optimized: Option<bool>,
    pub ena_support: Option<bool>,
    pub hypervisor: Option<String>,
    pub iam_instance_profile: Option<IamInstanceProfile>,
    pub image_id: Option<String>,
    pub instance_id: Option<String>,
    pub instance_lifecycle: Option<String>,
    pub instance_type: Option<String>,
    pub kernel_id: Option<String>,
    pub key_name: Option<String>,
    pub launch_time: Option<String>,
    pub monitoring: Option<Monitoring>,
    pub network_interfaces: Option<Vec<InstanceNetworkInterface>>,
    pub placement: Option<Placement>,
    pub platform: Option<String>,
    pub private_dns_name: Option<String>,
    pub private_ip_address: Option<String>,
    pub product_codes: Option<Vec<ProductCode>>,
    pub public_dns_name: Option<String>,
    pub public_ip_address: Option<String>,
    pub ramdisk_id: Option<String>,
    pub root_device_name: Option<String>,
    pub root_device_type: Option<String>,
    pub security_groups: Option<Vec<GroupIdentifier>>,
    pub source_dest_check: Option<bool>,
    pub spot_instance_request_id: Option<String>,
    pub sriov_net_support: Option<String>,
    pub state: Option<InstanceState>,
    pub state_reason: Option<StateReason>,
    pub state_transition_reason: Option<String>,
    pub subnet_id: Option<String>,
    pub tags: Option<Vec<Tag>>,
    pub virtualization_type: Option<String>,
    pub vpc_id: Option<String>,
}

Describes an instance.

Fields
ami_launch_index: Option<i64>

The AMI launch index, which can be used to find this instance in the launch group.

architecture: Option<String>

The architecture of the image.

block_device_mappings: Option<Vec<InstanceBlockDeviceMapping>>

Any block device mapping entries for the instance.

client_token: Option<String>

The idempotency token you provided when you launched the instance, if applicable.

ebs_optimized: Option<bool>

Indicates whether the instance is optimized for EBS I/O. This optimization provides dedicated throughput to Amazon EBS and an optimized configuration stack to provide optimal I/O performance. This optimization isn't available with all instance types. Additional usage charges apply when using an EBS Optimized instance.

ena_support: Option<bool>

Specifies whether enhanced networking with ENA is enabled.

hypervisor: Option<String>

The hypervisor type of the instance.

iam_instance_profile: Option<IamInstanceProfile>

The IAM instance profile associated with the instance, if applicable.

image_id: Option<String>

The ID of the AMI used to launch the instance.

instance_id: Option<String>

The ID of the instance.

instance_lifecycle: Option<String>

Indicates whether this is a Spot instance or a Scheduled Instance.

instance_type: Option<String>

The instance type.

kernel_id: Option<String>

The kernel associated with this instance, if applicable.

key_name: Option<String>

The name of the key pair, if this instance was launched with an associated key pair.

launch_time: Option<String>

The time the instance was launched.

monitoring: Option<Monitoring>

The monitoring for the instance.

network_interfaces: Option<Vec<InstanceNetworkInterface>>

[EC2-VPC] One or more network interfaces for the instance.

placement: Option<Placement>

The location where the instance launched, if applicable.

platform: Option<String>

The value is Windows for Windows instances; otherwise blank.

private_dns_name: Option<String>

(IPv4 only) The private DNS hostname name assigned to the instance. This DNS hostname can only be used inside the Amazon EC2 network. This name is not available until the instance enters the running state.

[EC2-VPC] The Amazon-provided DNS server will resolve Amazon-provided private DNS hostnames if you've enabled DNS resolution and DNS hostnames in your VPC. If you are not using the Amazon-provided DNS server in your VPC, your custom domain name servers must resolve the hostname as appropriate.

private_ip_address: Option<String>

The private IPv4 address assigned to the instance.

product_codes: Option<Vec<ProductCode>>

The product codes attached to this instance, if applicable.

public_dns_name: Option<String>

(IPv4 only) The public DNS name assigned to the instance. This name is not available until the instance enters the running state. For EC2-VPC, this name is only available if you've enabled DNS hostnames for your VPC.

public_ip_address: Option<String>

The public IPv4 address assigned to the instance, if applicable.

ramdisk_id: Option<String>

The RAM disk associated with this instance, if applicable.

root_device_name: Option<String>

The root device name (for example, /dev/sda1 or /dev/xvda).

root_device_type: Option<String>

The root device type used by the AMI. The AMI can use an EBS volume or an instance store volume.

security_groups: Option<Vec<GroupIdentifier>>

One or more security groups for the instance.

source_dest_check: Option<bool>

Specifies whether to enable an instance launched in a VPC to perform NAT. This controls whether source/destination checking is enabled on the instance. A value of true means checking is enabled, and false means checking is disabled. The value must be false for the instance to perform NAT. For more information, see NAT Instances in the Amazon Virtual Private Cloud User Guide.

spot_instance_request_id: Option<String>

If the request is a Spot instance request, the ID of the request.

sriov_net_support: Option<String>

Specifies whether enhanced networking with the Intel 82599 Virtual Function interface is enabled.

state: Option<InstanceState>

The current state of the instance.

state_reason: Option<StateReason>

The reason for the most recent state transition.

state_transition_reason: Option<String>

The reason for the most recent state transition. This might be an empty string.

subnet_id: Option<String>

[EC2-VPC] The ID of the subnet in which the instance is running.

tags: Option<Vec<Tag>>

Any tags assigned to the instance.

virtualization_type: Option<String>

The virtualization type of the instance.

vpc_id: Option<String>

[EC2-VPC] The ID of the VPC in which the instance is running.