Struct rusoto_ec2::InstanceNetworkInterfaceSpecification

pub struct InstanceNetworkInterfaceSpecification {
    pub associate_public_ip_address: Option<bool>,
    pub delete_on_termination: Option<bool>,
    pub description: Option<String>,
    pub device_index: Option<i64>,
    pub groups: Option<Vec<String>>,
    pub ipv_6_address_count: Option<i64>,
    pub ipv_6_addresses: Option<Vec<InstanceIpv6Address>>,
    pub network_interface_id: Option<String>,
    pub private_ip_address: Option<String>,
    pub private_ip_addresses: Option<Vec<PrivateIpAddressSpecification>>,
    pub secondary_private_ip_address_count: Option<i64>,
    pub subnet_id: Option<String>,
}

Describes a network interface.

Fields
associate_public_ip_address: Option<bool>

Indicates whether to assign a public IPv4 address to an instance you launch in a VPC. The public IP address can only be assigned to a network interface for eth0, and can only be assigned to a new network interface, not an existing one. You cannot specify more than one network interface in the request. If launching into a default subnet, the default value is true.

delete_on_termination: Option<bool>

If set to true, the interface is deleted when the instance is terminated. You can specify true only if creating a new network interface when launching an instance.

description: Option<String>

The description of the network interface. Applies only if creating a network interface when launching an instance.

device_index: Option<i64>

The index of the device on the instance for the network interface attachment. If you are specifying a network interface in a RunInstances request, you must provide the device index.

groups: Option<Vec<String>>

The IDs of the security groups for the network interface. Applies only if creating a network interface when launching an instance.

ipv_6_address_count: Option<i64>

A number of IPv6 addresses to assign to the network interface. Amazon EC2 chooses the IPv6 addresses from the range of the subnet. You cannot specify this option and the option to assign specific IPv6 addresses in the same request. You can specify this option if you've specified a minimum number of instances to launch.

ipv_6_addresses: Option<Vec<InstanceIpv6Address>>

One or more IPv6 addresses to assign to the network interface. You cannot specify this option and the option to assign a number of IPv6 addresses in the same request. You cannot specify this option if you've specified a minimum number of instances to launch.

network_interface_id: Option<String>

The ID of the network interface.

private_ip_address: Option<String>

The private IPv4 address of the network interface. Applies only if creating a network interface when launching an instance. You cannot specify this option if you're launching more than one instance in a RunInstances request.

private_ip_addresses: Option<Vec<PrivateIpAddressSpecification>>

One or more private IPv4 addresses to assign to the network interface. Only one private IPv4 address can be designated as primary. You cannot specify this option if you're launching more than one instance in a RunInstances request.

secondary_private_ip_address_count: Option<i64>

The number of secondary private IPv4 addresses. You can't specify this option and specify more than one private IP address using the private IP addresses option. You cannot specify this option if you're launching more than one instance in a RunInstances request.

subnet_id: Option<String>

The ID of the subnet associated with the network string. Applies only if creating a network interface when launching an instance.