Struct rusoto_ec2::IpPermission

pub struct IpPermission {
    pub from_port: Option<i64>,
    pub ip_protocol: Option<String>,
    pub ip_ranges: Option<Vec<IpRange>>,
    pub ipv_6_ranges: Option<Vec<Ipv6Range>>,
    pub prefix_list_ids: Option<Vec<PrefixListId>>,
    pub to_port: Option<i64>,
    pub user_id_group_pairs: Option<Vec<UserIdGroupPair>>,
}

Describes a security group rule.

Fields
from_port: Option<i64>

The start of port range for the TCP and UDP protocols, or an ICMP/ICMPv6 type number. A value of -1 indicates all ICMP/ICMPv6 types.

ip_protocol: Option<String>

The IP protocol name (tcp, udp, icmp) or number (see Protocol Numbers).

[EC2-VPC only] Use -1 to specify all protocols. When authorizing security group rules, specifying -1 or a protocol number other than tcp, udp, icmp, or 58 (ICMPv6) allows traffic on all ports, regardless of any port range you specify. For tcp, udp, and icmp, you must specify a port range. For 58 (ICMPv6), you can optionally specify a port range; if you don't, traffic for all types and codes is allowed when authorizing rules.

ip_ranges: Option<Vec<IpRange>>

One or more IPv4 ranges.

ipv_6_ranges: Option<Vec<Ipv6Range>>

[EC2-VPC only] One or more IPv6 ranges.

prefix_list_ids: Option<Vec<PrefixListId>>

(Valid for AuthorizeSecurityGroupEgress, RevokeSecurityGroupEgress and DescribeSecurityGroups only) One or more prefix list IDs for an AWS service. In an AuthorizeSecurityGroupEgress request, this is the AWS service that you want to access through a VPC endpoint from instances associated with the security group.

to_port: Option<i64>

The end of port range for the TCP and UDP protocols, or an ICMP/ICMPv6 code. A value of -1 indicates all ICMP/ICMPv6 codes for the specified ICMP type.

user_id_group_pairs: Option<Vec<UserIdGroupPair>>

One or more security group and AWS account ID pairs.