Struct rusoto_ec2::RegisterImageRequest

pub struct RegisterImageRequest {
    pub architecture: Option<String>,
    pub billing_products: Option<Vec<String>>,
    pub block_device_mappings: Option<Vec<BlockDeviceMapping>>,
    pub description: Option<String>,
    pub dry_run: Option<bool>,
    pub ena_support: Option<bool>,
    pub image_location: Option<String>,
    pub kernel_id: Option<String>,
    pub name: String,
    pub ramdisk_id: Option<String>,
    pub root_device_name: Option<String>,
    pub sriov_net_support: Option<String>,
    pub virtualization_type: Option<String>,
}

Contains the parameters for RegisterImage.

Fields
architecture: Option<String>

The architecture of the AMI.

Default: For Amazon EBS-backed AMIs, i386. For instance store-backed AMIs, the architecture specified in the manifest file.

billing_products: Option<Vec<String>>

The billing product codes. Your account must be authorized to specify billing product codes. Otherwise, you can use the AWS Marketplace to bill for the use of an AMI.

block_device_mappings: Option<Vec<BlockDeviceMapping>>

One or more block device mapping entries.

description: Option<String>

A description for your AMI.

dry_run: Option<bool>

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

ena_support: Option<bool>

Set to true to enable enhanced networking with ENA for the AMI and any instances that you launch from the AMI.

This option is supported only for HVM AMIs. Specifying this option with a PV AMI can make instances launched from the AMI unreachable.

image_location: Option<String>

The full path to your AMI manifest in Amazon S3 storage.

kernel_id: Option<String>

The ID of the kernel.

name: String

A name for your AMI.

Constraints: 3-128 alphanumeric characters, parentheses (()), square brackets ([]), spaces ( ), periods (.), slashes (/), dashes (-), single quotes ('), at-signs (@), or underscores(_)

ramdisk_id: Option<String>

The ID of the RAM disk.

root_device_name: Option<String>

The name of the root device (for example, /dev/sda1, or /dev/xvda).

sriov_net_support: Option<String>

Set to simple to enable enhanced networking with the Intel 82599 Virtual Function interface for the AMI and any instances that you launch from the AMI.

There is no way to disable sriovNetSupport at this time.

This option is supported only for HVM AMIs. Specifying this option with a PV AMI can make instances launched from the AMI unreachable.

virtualization_type: Option<String>

The type of virtualization.

Default: paravirtual