Struct rusoto_ec2::RequestSpotInstancesRequest

pub struct RequestSpotInstancesRequest {
    pub availability_zone_group: Option<String>,
    pub block_duration_minutes: Option<i64>,
    pub client_token: Option<String>,
    pub dry_run: Option<bool>,
    pub instance_count: Option<i64>,
    pub launch_group: Option<String>,
    pub launch_specification: Option<RequestSpotLaunchSpecification>,
    pub spot_price: String,
    pub type_: Option<String>,
    pub valid_from: Option<String>,
    pub valid_until: Option<String>,
}

Contains the parameters for RequestSpotInstances.

Fields
availability_zone_group: Option<String>

The user-specified name for a logical grouping of bids.

When you specify an Availability Zone group in a Spot Instance request, all Spot instances in the request are launched in the same Availability Zone. Instance proximity is maintained with this parameter, but the choice of Availability Zone is not. The group applies only to bids for Spot Instances of the same instance type. Any additional Spot instance requests that are specified with the same Availability Zone group name are launched in that same Availability Zone, as long as at least one instance from the group is still active.

If there is no active instance running in the Availability Zone group that you specify for a new Spot instance request (all instances are terminated, the bid is expired, or the bid falls below current market), then Amazon EC2 launches the instance in any Availability Zone where the constraint can be met. Consequently, the subsequent set of Spot instances could be placed in a different zone from the original request, even if you specified the same Availability Zone group.

Default: Instances are launched in any available Availability Zone.

block_duration_minutes: Option<i64>

The required duration for the Spot instances (also known as Spot blocks), in minutes. This value must be a multiple of 60 (60, 120, 180, 240, 300, or 360).

The duration period starts as soon as your Spot instance receives its instance ID. At the end of the duration period, Amazon EC2 marks the Spot instance for termination and provides a Spot instance termination notice, which gives the instance a two-minute warning before it terminates.

Note that you can't specify an Availability Zone group or a launch group if you specify a duration.

client_token: Option<String>

Unique, case-sensitive identifier that you provide to ensure the idempotency of the request. For more information, see How to Ensure Idempotency in the Amazon Elastic Compute Cloud User Guide.

dry_run: Option<bool>

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

instance_count: Option<i64>

The maximum number of Spot instances to launch.

Default: 1

launch_group: Option<String>

The instance launch group. Launch groups are Spot instances that launch together and terminate together.

Default: Instances are launched and terminated individually

launch_specification: Option<RequestSpotLaunchSpecification>

The launch specification.

spot_price: String

The maximum hourly price (bid) for any Spot instance launched to fulfill the request.

type_: Option<String>

The Spot instance request type.

Default: one-time

valid_from: Option<String>

The start date of the request. If this is a one-time request, the request becomes active at this date and time and remains active until all instances launch, the request expires, or the request is canceled. If the request is persistent, the request becomes active at this date and time and remains active until it expires or is canceled.

Default: The request is effective indefinitely.

valid_until: Option<String>

The end date of the request. If this is a one-time request, the request remains active until all instances launch, the request is canceled, or this date is reached. If the request is persistent, it remains active until it is canceled or this date and time is reached.

Default: The request is effective indefinitely.