Struct rusoto_ec2::ReservedInstancesModification

pub struct ReservedInstancesModification {
    pub client_token: Option<String>,
    pub create_date: Option<String>,
    pub effective_date: Option<String>,
    pub modification_results: Option<Vec<ReservedInstancesModificationResult>>,
    pub reserved_instances_ids: Option<Vec<ReservedInstancesId>>,
    pub reserved_instances_modification_id: Option<String>,
    pub status: Option<String>,
    pub status_message: Option<String>,
    pub update_date: Option<String>,
}

Describes a Reserved Instance modification.

Fields
client_token: Option<String>

A unique, case-sensitive key supplied by the client to ensure that the request is idempotent. For more information, see Ensuring Idempotency.

create_date: Option<String>

The time when the modification request was created.

effective_date: Option<String>

The time for the modification to become effective.

modification_results: Option<Vec<ReservedInstancesModificationResult>>

Contains target configurations along with their corresponding new Reserved Instance IDs.

reserved_instances_ids: Option<Vec<ReservedInstancesId>>

The IDs of one or more Reserved Instances.

reserved_instances_modification_id: Option<String>

A unique ID for the Reserved Instance modification.

status: Option<String>

The status of the Reserved Instances modification request.

status_message: Option<String>

The reason for the status.

update_date: Option<String>

The time when the modification request was last updated.