Struct rusoto_ec2::ReservedInstancesOffering

pub struct ReservedInstancesOffering {
    pub availability_zone: Option<String>,
    pub currency_code: Option<String>,
    pub duration: Option<i64>,
    pub fixed_price: Option<f32>,
    pub instance_tenancy: Option<String>,
    pub instance_type: Option<String>,
    pub marketplace: Option<bool>,
    pub offering_class: Option<String>,
    pub offering_type: Option<String>,
    pub pricing_details: Option<Vec<PricingDetail>>,
    pub product_description: Option<String>,
    pub recurring_charges: Option<Vec<RecurringCharge>>,
    pub reserved_instances_offering_id: Option<String>,
    pub scope: Option<String>,
    pub usage_price: Option<f32>,
}

Describes a Reserved Instance offering.

Fields
availability_zone: Option<String>

The Availability Zone in which the Reserved Instance can be used.

currency_code: Option<String>

The currency of the Reserved Instance offering you are purchasing. It's specified using ISO 4217 standard currency codes. At this time, the only supported currency is USD.

duration: Option<i64>

The duration of the Reserved Instance, in seconds.

fixed_price: Option<f32>

The purchase price of the Reserved Instance.

instance_tenancy: Option<String>

The tenancy of the instance.

instance_type: Option<String>

The instance type on which the Reserved Instance can be used.

marketplace: Option<bool>

Indicates whether the offering is available through the Reserved Instance Marketplace (resale) or AWS. If it's a Reserved Instance Marketplace offering, this is true.

offering_class: Option<String>

If convertible it can be exchanged for Reserved Instances of the same or higher monetary value, with different configurations. If standard, it is not possible to perform an exchange.

offering_type: Option<String>

The Reserved Instance offering type.

pricing_details: Option<Vec<PricingDetail>>

The pricing details of the Reserved Instance offering.

product_description: Option<String>

The Reserved Instance product platform description.

recurring_charges: Option<Vec<RecurringCharge>>

The recurring charge tag assigned to the resource.

reserved_instances_offering_id: Option<String>

The ID of the Reserved Instance offering. This is the offering ID used in GetReservedInstancesExchangeQuote to confirm that an exchange can be made.

scope: Option<String>

Whether the Reserved Instance is applied to instances in a region or an Availability Zone.

usage_price: Option<f32>

The usage price of the Reserved Instance, per hour.