Struct rusoto_ec2::Route

pub struct Route {
    pub destination_cidr_block: Option<String>,
    pub destination_ipv_6_cidr_block: Option<String>,
    pub destination_prefix_list_id: Option<String>,
    pub egress_only_internet_gateway_id: Option<String>,
    pub gateway_id: Option<String>,
    pub instance_id: Option<String>,
    pub instance_owner_id: Option<String>,
    pub nat_gateway_id: Option<String>,
    pub network_interface_id: Option<String>,
    pub origin: Option<String>,
    pub state: Option<String>,
    pub vpc_peering_connection_id: Option<String>,
}

Describes a route in a route table.

Fields
destination_cidr_block: Option<String>

The IPv4 CIDR block used for the destination match.

destination_ipv_6_cidr_block: Option<String>

The IPv6 CIDR block used for the destination match.

destination_prefix_list_id: Option<String>

The prefix of the AWS service.

egress_only_internet_gateway_id: Option<String>

The ID of the egress-only Internet gateway.

gateway_id: Option<String>

The ID of a gateway attached to your VPC.

instance_id: Option<String>

The ID of a NAT instance in your VPC.

instance_owner_id: Option<String>

The AWS account ID of the owner of the instance.

nat_gateway_id: Option<String>

The ID of a NAT gateway.

network_interface_id: Option<String>

The ID of the network interface.

origin: Option<String>

Describes how the route was created.

CreateRouteTable - The route was automatically created when the route table was created.

CreateRoute - The route was manually added to the route table.

EnableVgwRoutePropagation - The route was propagated by route propagation.

state: Option<String>

The state of the route. The blackhole state indicates that the route's target isn't available (for example, the specified gateway isn't attached to the VPC, or the specified NAT instance has been terminated).

vpc_peering_connection_id: Option<String>

The ID of the VPC peering connection.