Struct rusoto_ec2::RunInstancesRequest

pub struct RunInstancesRequest {
    pub additional_info: Option<String>,
    pub block_device_mappings: Option<Vec<BlockDeviceMapping>>,
    pub client_token: Option<String>,
    pub disable_api_termination: Option<bool>,
    pub dry_run: Option<bool>,
    pub ebs_optimized: Option<bool>,
    pub iam_instance_profile: Option<IamInstanceProfileSpecification>,
    pub image_id: String,
    pub instance_initiated_shutdown_behavior: Option<String>,
    pub instance_type: Option<String>,
    pub ipv_6_address_count: Option<i64>,
    pub ipv_6_addresses: Option<Vec<InstanceIpv6Address>>,
    pub kernel_id: Option<String>,
    pub key_name: Option<String>,
    pub max_count: i64,
    pub min_count: i64,
    pub monitoring: Option<RunInstancesMonitoringEnabled>,
    pub network_interfaces: Option<Vec<InstanceNetworkInterfaceSpecification>>,
    pub placement: Option<Placement>,
    pub private_ip_address: Option<String>,
    pub ramdisk_id: Option<String>,
    pub security_group_ids: Option<Vec<String>>,
    pub security_groups: Option<Vec<String>>,
    pub subnet_id: Option<String>,
    pub tag_specifications: Option<Vec<TagSpecification>>,
    pub user_data: Option<String>,
}

Contains the parameters for RunInstances.

Fields
additional_info: Option<String>

Reserved.

block_device_mappings: Option<Vec<BlockDeviceMapping>>

The block device mapping.

Supplying both a snapshot ID and an encryption value as arguments for block-device mapping results in an error. This is because only blank volumes can be encrypted on start, and these are not created from a snapshot. If a snapshot is the basis for the volume, it contains data by definition and its encryption status cannot be changed using this action.

client_token: Option<String>

Unique, case-sensitive identifier you provide to ensure the idempotency of the request. For more information, see Ensuring Idempotency.

Constraints: Maximum 64 ASCII characters

disable_api_termination: Option<bool>

If you set this parameter to true, you can't terminate the instance using the Amazon EC2 console, CLI, or API; otherwise, you can. To change this attribute to false after launch, use ModifyInstanceAttribute. Alternatively, if you set InstanceInitiatedShutdownBehavior to terminate, you can terminate the instance by running the shutdown command from the instance.

Default: false

dry_run: Option<bool>

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

ebs_optimized: Option<bool>

Indicates whether the instance is optimized for EBS I/O. This optimization provides dedicated throughput to Amazon EBS and an optimized configuration stack to provide optimal EBS I/O performance. This optimization isn't available with all instance types. Additional usage charges apply when using an EBS-optimized instance.

Default: false

iam_instance_profile: Option<IamInstanceProfileSpecification>

The IAM instance profile.

image_id: String

The ID of the AMI, which you can get by calling DescribeImages.

instance_initiated_shutdown_behavior: Option<String>

Indicates whether an instance stops or terminates when you initiate shutdown from the instance (using the operating system command for system shutdown).

Default: stop

instance_type: Option<String>

The instance type. For more information, see Instance Types in the Amazon Elastic Compute Cloud User Guide.

Default: m1.small

ipv_6_address_count: Option<i64>

[EC2-VPC] A number of IPv6 addresses to associate with the primary network interface. Amazon EC2 chooses the IPv6 addresses from the range of your subnet. You cannot specify this option and the option to assign specific IPv6 addresses in the same request. You can specify this option if you've specified a minimum number of instances to launch.

ipv_6_addresses: Option<Vec<InstanceIpv6Address>>

[EC2-VPC] Specify one or more IPv6 addresses from the range of the subnet to associate with the primary network interface. You cannot specify this option and the option to assign a number of IPv6 addresses in the same request. You cannot specify this option if you've specified a minimum number of instances to launch.

kernel_id: Option<String>

The ID of the kernel.

We recommend that you use PV-GRUB instead of kernels and RAM disks. For more information, see PV-GRUB in the Amazon Elastic Compute Cloud User Guide.

key_name: Option<String>

The name of the key pair. You can create a key pair using CreateKeyPair or ImportKeyPair.

If you do not specify a key pair, you can't connect to the instance unless you choose an AMI that is configured to allow users another way to log in.

max_count: i64

The maximum number of instances to launch. If you specify more instances than Amazon EC2 can launch in the target Availability Zone, Amazon EC2 launches the largest possible number of instances above MinCount.

Constraints: Between 1 and the maximum number you're allowed for the specified instance type. For more information about the default limits, and how to request an increase, see How many instances can I run in Amazon EC2 in the Amazon EC2 FAQ.

min_count: i64

The minimum number of instances to launch. If you specify a minimum that is more instances than Amazon EC2 can launch in the target Availability Zone, Amazon EC2 launches no instances.

Constraints: Between 1 and the maximum number you're allowed for the specified instance type. For more information about the default limits, and how to request an increase, see How many instances can I run in Amazon EC2 in the Amazon EC2 General FAQ.

monitoring: Option<RunInstancesMonitoringEnabled>

The monitoring for the instance.

network_interfaces: Option<Vec<InstanceNetworkInterfaceSpecification>>

One or more network interfaces.

placement: Option<Placement>

The placement for the instance.

private_ip_address: Option<String>

[EC2-VPC] The primary IPv4 address. You must specify a value from the IPv4 address range of the subnet.

Only one private IP address can be designated as primary. You can't specify this option if you've specified the option to designate a private IP address as the primary IP address in a network interface specification. You cannot specify this option if you're launching more than one instance in the request.

ramdisk_id: Option<String>

The ID of the RAM disk.

We recommend that you use PV-GRUB instead of kernels and RAM disks. For more information, see PV-GRUB in the Amazon Elastic Compute Cloud User Guide.

security_group_ids: Option<Vec<String>>

One or more security group IDs. You can create a security group using CreateSecurityGroup.

Default: Amazon EC2 uses the default security group.

security_groups: Option<Vec<String>>

[EC2-Classic, default VPC] One or more security group names. For a nondefault VPC, you must use security group IDs instead.

Default: Amazon EC2 uses the default security group.

subnet_id: Option<String>

[EC2-VPC] The ID of the subnet to launch the instance into.

tag_specifications: Option<Vec<TagSpecification>>

The tags to apply to the resources during launch. You can tag instances and volumes. The specified tags are applied to all instances or volumes that are created during launch.

user_data: Option<String>

The user data to make available to the instance. For more information, see Running Commands on Your Linux Instance at Launch (Linux) and Adding User Data (Windows). If you are using an AWS SDK or command line tool, Base64-encoding is performed for you, and you can load the text from a file. Otherwise, you must provide Base64-encoded text.