Struct rusoto_ec2::ScheduledInstance

pub struct ScheduledInstance {
    pub availability_zone: Option<String>,
    pub create_date: Option<String>,
    pub hourly_price: Option<String>,
    pub instance_count: Option<i64>,
    pub instance_type: Option<String>,
    pub network_platform: Option<String>,
    pub next_slot_start_time: Option<String>,
    pub platform: Option<String>,
    pub previous_slot_end_time: Option<String>,
    pub recurrence: Option<ScheduledInstanceRecurrence>,
    pub scheduled_instance_id: Option<String>,
    pub slot_duration_in_hours: Option<i64>,
    pub term_end_date: Option<String>,
    pub term_start_date: Option<String>,
    pub total_scheduled_instance_hours: Option<i64>,
}

Describes a Scheduled Instance.

Fields
availability_zone: Option<String>

The Availability Zone.

create_date: Option<String>

The date when the Scheduled Instance was purchased.

hourly_price: Option<String>

The hourly price for a single instance.

instance_count: Option<i64>

The number of instances.

instance_type: Option<String>

The instance type.

network_platform: Option<String>

The network platform (EC2-Classic or EC2-VPC).

next_slot_start_time: Option<String>

The time for the next schedule to start.

platform: Option<String>

The platform (Linux/UNIX or Windows).

previous_slot_end_time: Option<String>

The time that the previous schedule ended or will end.

recurrence: Option<ScheduledInstanceRecurrence>

The schedule recurrence.

scheduled_instance_id: Option<String>

The Scheduled Instance ID.

slot_duration_in_hours: Option<i64>

The number of hours in the schedule.

term_end_date: Option<String>

The end date for the Scheduled Instance.

term_start_date: Option<String>

The start date for the Scheduled Instance.

total_scheduled_instance_hours: Option<i64>

The total number of hours for a single instance for the entire term.