Struct rusoto_ec2::Snapshot

pub struct Snapshot {
    pub data_encryption_key_id: Option<String>,
    pub description: Option<String>,
    pub encrypted: Option<bool>,
    pub kms_key_id: Option<String>,
    pub owner_alias: Option<String>,
    pub owner_id: Option<String>,
    pub progress: Option<String>,
    pub snapshot_id: Option<String>,
    pub start_time: Option<String>,
    pub state: Option<String>,
    pub state_message: Option<String>,
    pub tags: Option<Vec<Tag>>,
    pub volume_id: Option<String>,
    pub volume_size: Option<i64>,
}

Describes a snapshot.

Fields
data_encryption_key_id: Option<String>

The data encryption key identifier for the snapshot. This value is a unique identifier that corresponds to the data encryption key that was used to encrypt the original volume or snapshot copy. Because data encryption keys are inherited by volumes created from snapshots, and vice versa, if snapshots share the same data encryption key identifier, then they belong to the same volume/snapshot lineage. This parameter is only returned by the DescribeSnapshots API operation.

description: Option<String>

The description for the snapshot.

encrypted: Option<bool>

Indicates whether the snapshot is encrypted.

kms_key_id: Option<String>

The full ARN of the AWS Key Management Service (AWS KMS) customer master key (CMK) that was used to protect the volume encryption key for the parent volume.

owner_alias: Option<String>

Value from an Amazon-maintained list (amazon | aws-marketplace | microsoft) of snapshot owners. Not to be confused with the user-configured AWS account alias, which is set from the IAM console.

owner_id: Option<String>

The AWS account ID of the EBS snapshot owner.

progress: Option<String>

The progress of the snapshot, as a percentage.

snapshot_id: Option<String>

The ID of the snapshot. Each snapshot receives a unique identifier when it is created.

start_time: Option<String>

The time stamp when the snapshot was initiated.

state: Option<String>

The snapshot state.

state_message: Option<String>

Encrypted Amazon EBS snapshots are copied asynchronously. If a snapshot copy operation fails (for example, if the proper AWS Key Management Service (AWS KMS) permissions are not obtained) this field displays error state details to help you diagnose why the error occurred. This parameter is only returned by the DescribeSnapshots API operation.

tags: Option<Vec<Tag>>

Any tags assigned to the snapshot.

volume_id: Option<String>

The ID of the volume that was used to create the snapshot. Snapshots created by the CopySnapshot action have an arbitrary volume ID that should not be used for any purpose.

volume_size: Option<i64>

The size of the volume, in GiB.