Struct rusoto_ec2::SpotFleetLaunchSpecification

pub struct SpotFleetLaunchSpecification {
    pub addressing_type: Option<String>,
    pub block_device_mappings: Option<Vec<BlockDeviceMapping>>,
    pub ebs_optimized: Option<bool>,
    pub iam_instance_profile: Option<IamInstanceProfileSpecification>,
    pub image_id: Option<String>,
    pub instance_type: Option<String>,
    pub kernel_id: Option<String>,
    pub key_name: Option<String>,
    pub monitoring: Option<SpotFleetMonitoring>,
    pub network_interfaces: Option<Vec<InstanceNetworkInterfaceSpecification>>,
    pub placement: Option<SpotPlacement>,
    pub ramdisk_id: Option<String>,
    pub security_groups: Option<Vec<GroupIdentifier>>,
    pub spot_price: Option<String>,
    pub subnet_id: Option<String>,
    pub user_data: Option<String>,
    pub weighted_capacity: Option<f64>,
}

Describes the launch specification for one or more Spot instances.

Fields
addressing_type: Option<String>

Deprecated.

block_device_mappings: Option<Vec<BlockDeviceMapping>>

One or more block device mapping entries.

ebs_optimized: Option<bool>

Indicates whether the instances are optimized for EBS I/O. This optimization provides dedicated throughput to Amazon EBS and an optimized configuration stack to provide optimal EBS I/O performance. This optimization isn't available with all instance types. Additional usage charges apply when using an EBS Optimized instance.

Default: false

iam_instance_profile: Option<IamInstanceProfileSpecification>

The IAM instance profile.

image_id: Option<String>

The ID of the AMI.

instance_type: Option<String>

The instance type. Note that T2 and HS1 instance types are not supported.

kernel_id: Option<String>

The ID of the kernel.

key_name: Option<String>

The name of the key pair.

monitoring: Option<SpotFleetMonitoring>

Enable or disable monitoring for the instances.

network_interfaces: Option<Vec<InstanceNetworkInterfaceSpecification>>

One or more network interfaces. If you specify a network interface, you must specify subnet IDs and security group IDs using the network interface.

placement: Option<SpotPlacement>

The placement information.

ramdisk_id: Option<String>

The ID of the RAM disk.

security_groups: Option<Vec<GroupIdentifier>>

One or more security groups. When requesting instances in a VPC, you must specify the IDs of the security groups. When requesting instances in EC2-Classic, you can specify the names or the IDs of the security groups.

spot_price: Option<String>

The bid price per unit hour for the specified instance type. If this value is not specified, the default is the Spot bid price specified for the fleet. To determine the bid price per unit hour, divide the Spot bid price by the value of WeightedCapacity.

subnet_id: Option<String>

The ID of the subnet in which to launch the instances. To specify multiple subnets, separate them using commas; for example, "subnet-a61dafcf, subnet-65ea5f08".

user_data: Option<String>

The user data to make available to the instances. If you are using an AWS SDK or command line tool, Base64-encoding is performed for you, and you can load the text from a file. Otherwise, you must provide Base64-encoded text.

weighted_capacity: Option<f64>

The number of units provided by the specified instance type. These are the same units that you chose to set the target capacity in terms (instances or a performance characteristic such as vCPUs, memory, or I/O).

If the target capacity divided by this value is not a whole number, we round the number of instances to the next whole number. If this value is not specified, the default is 1.