Struct rusoto_ec2::SpotFleetRequestConfigData

pub struct SpotFleetRequestConfigData {
    pub allocation_strategy: Option<String>,
    pub client_token: Option<String>,
    pub excess_capacity_termination_policy: Option<String>,
    pub fulfilled_capacity: Option<f64>,
    pub iam_fleet_role: String,
    pub launch_specifications: Vec<SpotFleetLaunchSpecification>,
    pub replace_unhealthy_instances: Option<bool>,
    pub spot_price: String,
    pub target_capacity: i64,
    pub terminate_instances_with_expiration: Option<bool>,
    pub type_: Option<String>,
    pub valid_from: Option<String>,
    pub valid_until: Option<String>,
}

Describes the configuration of a Spot fleet request.

Fields
allocation_strategy: Option<String>

Indicates how to allocate the target capacity across the Spot pools specified by the Spot fleet request. The default is lowestPrice.

client_token: Option<String>

A unique, case-sensitive identifier you provide to ensure idempotency of your listings. This helps avoid duplicate listings. For more information, see Ensuring Idempotency.

excess_capacity_termination_policy: Option<String>

Indicates whether running Spot instances should be terminated if the target capacity of the Spot fleet request is decreased below the current size of the Spot fleet.

fulfilled_capacity: Option<f64>

The number of units fulfilled by this request compared to the set target capacity.

iam_fleet_role: String

Grants the Spot fleet permission to terminate Spot instances on your behalf when you cancel its Spot fleet request using CancelSpotFleetRequests or when the Spot fleet request expires, if you set terminateInstancesWithExpiration.

launch_specifications: Vec<SpotFleetLaunchSpecification>

Information about the launch specifications for the Spot fleet request.

replace_unhealthy_instances: Option<bool>

Indicates whether Spot fleet should replace unhealthy instances.

spot_price: String

The bid price per unit hour.

target_capacity: i64

The number of units to request. You can choose to set the target capacity in terms of instances or a performance characteristic that is important to your application workload, such as vCPUs, memory, or I/O.

terminate_instances_with_expiration: Option<bool>

Indicates whether running Spot instances should be terminated when the Spot fleet request expires.

type_: Option<String>

The type of request. Indicates whether the fleet will only request the target capacity or also attempt to maintain it. When you request a certain target capacity, the fleet will only place the required bids. It will not attempt to replenish Spot instances if capacity is diminished, nor will it submit bids in alternative Spot pools if capacity is not available. When you want to maintain a certain target capacity, fleet will place the required bids to meet this target capacity. It will also automatically replenish any interrupted instances. Default: maintain.

valid_from: Option<String>

The start date and time of the request, in UTC format (for example, YYYY-MM-DDTHH:MM:SSZ). The default is to start fulfilling the request immediately.

valid_until: Option<String>

The end date and time of the request, in UTC format (for example, YYYY-MM-DDTHH:MM:SSZ). At this point, no new Spot instance requests are placed or enabled to fulfill the request.