Struct rusoto_ec2::Subnet

pub struct Subnet {
    pub assign_ipv_6_address_on_creation: Option<bool>,
    pub availability_zone: Option<String>,
    pub available_ip_address_count: Option<i64>,
    pub cidr_block: Option<String>,
    pub default_for_az: Option<bool>,
    pub ipv_6_cidr_block_association_set: Option<Vec<SubnetIpv6CidrBlockAssociation>>,
    pub map_public_ip_on_launch: Option<bool>,
    pub state: Option<String>,
    pub subnet_id: Option<String>,
    pub tags: Option<Vec<Tag>>,
    pub vpc_id: Option<String>,
}

Describes a subnet.

Fields
assign_ipv_6_address_on_creation: Option<bool>

Indicates whether a network interface created in this subnet (including a network interface created by RunInstances) receives an IPv6 address.

availability_zone: Option<String>

The Availability Zone of the subnet.

available_ip_address_count: Option<i64>

The number of unused private IPv4 addresses in the subnet. Note that the IPv4 addresses for any stopped instances are considered unavailable.

cidr_block: Option<String>

The IPv4 CIDR block assigned to the subnet.

default_for_az: Option<bool>

Indicates whether this is the default subnet for the Availability Zone.

ipv_6_cidr_block_association_set: Option<Vec<SubnetIpv6CidrBlockAssociation>>

Information about the IPv6 CIDR blocks associated with the subnet.

map_public_ip_on_launch: Option<bool>

Indicates whether instances launched in this subnet receive a public IPv4 address.

state: Option<String>

The current state of the subnet.

subnet_id: Option<String>

The ID of the subnet.

tags: Option<Vec<Tag>>

Any tags assigned to the subnet.

vpc_id: Option<String>

The ID of the VPC the subnet is in.