Struct rusoto_ec2::VpnTunnelOptionsSpecification

pub struct VpnTunnelOptionsSpecification {
    pub pre_shared_key: Option<String>,
    pub tunnel_inside_cidr: Option<String>,
}

The tunnel options for a VPN connection.

Fields
pre_shared_key: Option<String>

The pre-shared key (PSK) to establish initial authentication between the virtual private gateway and customer gateway.

Constraints: Allowed characters are alphanumeric characters and ._. Must be between 8 and 64 characters in length and cannot start with zero (0).

tunnel_inside_cidr: Option<String>

The range of inside IP addresses for the tunnel. Any specified CIDR blocks must be unique across all VPN connections that use the same virtual private gateway.

Constraints: A size /30 CIDR block from the 169.254.0.0/16 range. The following CIDR blocks are reserved and cannot be used:

169.254.0.0/30

169.254.1.0/30

169.254.2.0/30

169.254.3.0/30

169.254.4.0/30

169.254.5.0/30

169.254.169.252/30