Struct rusoto_ec2::ModifyInstanceAttributeRequest

pub struct ModifyInstanceAttributeRequest {
    pub attribute: Option<String>,
    pub block_device_mappings: Option<Vec<InstanceBlockDeviceMappingSpecification>>,
    pub disable_api_termination: Option<AttributeBooleanValue>,
    pub dry_run: Option<bool>,
    pub ebs_optimized: Option<AttributeBooleanValue>,
    pub ena_support: Option<AttributeBooleanValue>,
    pub groups: Option<Vec<String>>,
    pub instance_id: String,
    pub instance_initiated_shutdown_behavior: Option<AttributeValue>,
    pub instance_type: Option<AttributeValue>,
    pub kernel: Option<AttributeValue>,
    pub ramdisk: Option<AttributeValue>,
    pub source_dest_check: Option<AttributeBooleanValue>,
    pub sriov_net_support: Option<AttributeValue>,
    pub user_data: Option<BlobAttributeValue>,
    pub value: Option<String>,
}

Contains the parameters for ModifyInstanceAttribute.

Fields
attribute: Option<String>

The name of the attribute.

block_device_mappings: Option<Vec<InstanceBlockDeviceMappingSpecification>>

Modifies the DeleteOnTermination attribute for volumes that are currently attached. The volume must be owned by the caller. If no value is specified for DeleteOnTermination, the default is true and the volume is deleted when the instance is terminated.

To add instance store volumes to an Amazon EBS-backed instance, you must add them when you launch the instance. For more information, see Updating the Block Device Mapping when Launching an Instance in the Amazon Elastic Compute Cloud User Guide.

disable_api_termination: Option<AttributeBooleanValue>

If the value is true, you can't terminate the instance using the Amazon EC2 console, CLI, or API; otherwise, you can. You cannot use this parameter for Spot Instances.

dry_run: Option<bool>

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

ebs_optimized: Option<AttributeBooleanValue>

Specifies whether the instance is optimized for Amazon EBS I/O. This optimization provides dedicated throughput to Amazon EBS and an optimized configuration stack to provide optimal EBS I/O performance. This optimization isn't available with all instance types. Additional usage charges apply when using an EBS Optimized instance.

ena_support: Option<AttributeBooleanValue>

Set to true to enable enhanced networking with ENA for the instance.

This option is supported only for HVM instances. Specifying this option with a PV instance can make it unreachable.

groups: Option<Vec<String>>

[EC2-VPC] Changes the security groups of the instance. You must specify at least one security group, even if it's just the default security group for the VPC. You must specify the security group ID, not the security group name.

instance_id: String

The ID of the instance.

instance_initiated_shutdown_behavior: Option<AttributeValue>

Specifies whether an instance stops or terminates when you initiate shutdown from the instance (using the operating system command for system shutdown).

instance_type: Option<AttributeValue>

Changes the instance type to the specified value. For more information, see Instance Types. If the instance type is not valid, the error returned is InvalidInstanceAttributeValue.

kernel: Option<AttributeValue>

Changes the instance's kernel to the specified value. We recommend that you use PV-GRUB instead of kernels and RAM disks. For more information, see PV-GRUB.

ramdisk: Option<AttributeValue>

Changes the instance's RAM disk to the specified value. We recommend that you use PV-GRUB instead of kernels and RAM disks. For more information, see PV-GRUB.

source_dest_check: Option<AttributeBooleanValue>

Specifies whether source/destination checking is enabled. A value of true means that checking is enabled, and false means that checking is disabled. This value must be false for a NAT instance to perform NAT.

sriov_net_support: Option<AttributeValue>

Set to simple to enable enhanced networking with the Intel 82599 Virtual Function interface for the instance.

There is no way to disable enhanced networking with the Intel 82599 Virtual Function interface at this time.

This option is supported only for HVM instances. Specifying this option with a PV instance can make it unreachable.

user_data: Option<BlobAttributeValue>

Changes the instance's user data to the specified value. If you are using an AWS SDK or command line tool, base64-encoding is performed for you, and you can load the text from a file. Otherwise, you must provide base64-encoded text.

value: Option<String>

A new value for the attribute. Use only with the kernel, ramdisk, userData, disableApiTermination, or instanceInitiatedShutdownBehavior attribute.