Struct rusoto_ec2::SpotFleetRequestConfigData

pub struct SpotFleetRequestConfigData {
    pub allocation_strategy: Option<String>,
    pub client_token: Option<String>,
    pub excess_capacity_termination_policy: Option<String>,
    pub fulfilled_capacity: Option<f64>,
    pub iam_fleet_role: String,
    pub instance_interruption_behavior: Option<String>,
    pub launch_specifications: Option<Vec<SpotFleetLaunchSpecification>>,
    pub launch_template_configs: Option<Vec<LaunchTemplateConfig>>,
    pub load_balancers_config: Option<LoadBalancersConfig>,
    pub replace_unhealthy_instances: Option<bool>,
    pub spot_price: Option<String>,
    pub target_capacity: i64,
    pub terminate_instances_with_expiration: Option<bool>,
    pub type_: Option<String>,
    pub valid_from: Option<String>,
    pub valid_until: Option<String>,
}

Describes the configuration of a Spot Fleet request.

Fields
allocation_strategy: Option<String>

Indicates how to allocate the target capacity across the Spot pools specified by the Spot Fleet request. The default is lowestPrice.

client_token: Option<String>

A unique, case-sensitive identifier you provide to ensure idempotency of your listings. This helps avoid duplicate listings. For more information, see Ensuring Idempotency.

excess_capacity_termination_policy: Option<String>

Indicates whether running Spot Instances should be terminated if the target capacity of the Spot Fleet request is decreased below the current size of the Spot Fleet.

fulfilled_capacity: Option<f64>

The number of units fulfilled by this request compared to the set target capacity.

iam_fleet_role: String

Grants the Spot Fleet permission to terminate Spot Instances on your behalf when you cancel its Spot Fleet request using CancelSpotFleetRequests or when the Spot Fleet request expires, if you set terminateInstancesWithExpiration.

instance_interruption_behavior: Option<String>

The behavior when a Spot Instance is interrupted. The default is terminate.

launch_specifications: Option<Vec<SpotFleetLaunchSpecification>>

The launch specifications for the Spot Fleet request.

launch_template_configs: Option<Vec<LaunchTemplateConfig>>

The launch template and overrides.

load_balancers_config: Option<LoadBalancersConfig>

One or more Classic Load Balancers and target groups to attach to the Spot Fleet request. Spot Fleet registers the running Spot Instances with the specified Classic Load Balancers and target groups.

With Network Load Balancers, Spot Fleet cannot register instances that have the following instance types: C1, CC1, CC2, CG1, CG2, CR1, CS1, G1, G2, HI1, HS1, M1, M2, M3, and T1.

replace_unhealthy_instances: Option<bool>

Indicates whether Spot Fleet should replace unhealthy instances.

spot_price: Option<String>

The maximum price per unit hour that you are willing to pay for a Spot Instance. The default is the On-Demand price.

target_capacity: i64

The number of units to request. You can choose to set the target capacity in terms of instances or a performance characteristic that is important to your application workload, such as vCPUs, memory, or I/O. If the request type is maintain, you can specify a target capacity of 0 and add capacity later.

terminate_instances_with_expiration: Option<bool>

Indicates whether running Spot Instances should be terminated when the Spot Fleet request expires.

type_: Option<String>

The type of request. Indicates whether the fleet will only request the target capacity or also attempt to maintain it. When you request a certain target capacity, the fleet will only place the required requests. It will not attempt to replenish Spot Instances if capacity is diminished, nor will it submit requests in alternative Spot pools if capacity is not available. When you want to maintain a certain target capacity, fleet will place the required requests to meet this target capacity. It will also automatically replenish any interrupted instances. Default: maintain.

valid_from: Option<String>

The start date and time of the request, in UTC format (for example, YYYY-MM-DDTHH:MM:SSZ). The default is to start fulfilling the request immediately.

valid_until: Option<String>

The end date and time of the request, in UTC format (for example, YYYY-MM-DDTHH:MM:SSZ). At this point, no new Spot Instance requests are placed or able to fulfill the request. The default end date is 7 days from the current date.