Struct rusoto_ec2::SpotInstanceRequest

pub struct SpotInstanceRequest {
    pub actual_block_hourly_price: Option<String>,
    pub availability_zone_group: Option<String>,
    pub block_duration_minutes: Option<i64>,
    pub create_time: Option<String>,
    pub fault: Option<SpotInstanceStateFault>,
    pub instance_id: Option<String>,
    pub instance_interruption_behavior: Option<String>,
    pub launch_group: Option<String>,
    pub launch_specification: Option<LaunchSpecification>,
    pub launched_availability_zone: Option<String>,
    pub product_description: Option<String>,
    pub spot_instance_request_id: Option<String>,
    pub spot_price: Option<String>,
    pub state: Option<String>,
    pub status: Option<SpotInstanceStatus>,
    pub tags: Option<Vec<Tag>>,
    pub type_: Option<String>,
    pub valid_from: Option<String>,
    pub valid_until: Option<String>,
}

Describes a Spot Instance request.

Fields
actual_block_hourly_price: Option<String>

If you specified a duration and your Spot Instance request was fulfilled, this is the fixed hourly price in effect for the Spot Instance while it runs.

availability_zone_group: Option<String>

The Availability Zone group. If you specify the same Availability Zone group for all Spot Instance requests, all Spot Instances are launched in the same Availability Zone.

block_duration_minutes: Option<i64>

The duration for the Spot Instance, in minutes.

create_time: Option<String>

The date and time when the Spot Instance request was created, in UTC format (for example, YYYY-MM-DDTHH:MM:SSZ).

fault: Option<SpotInstanceStateFault>

The fault codes for the Spot Instance request, if any.

instance_id: Option<String>

The instance ID, if an instance has been launched to fulfill the Spot Instance request.

instance_interruption_behavior: Option<String>

The behavior when a Spot Instance is interrupted.

launch_group: Option<String>

The instance launch group. Launch groups are Spot Instances that launch together and terminate together.

launch_specification: Option<LaunchSpecification>

Additional information for launching instances.

launched_availability_zone: Option<String>

The Availability Zone in which the request is launched.

product_description: Option<String>

The product description associated with the Spot Instance.

spot_instance_request_id: Option<String>

The ID of the Spot Instance request.

spot_price: Option<String>

The maximum price per hour that you are willing to pay for a Spot Instance.

state: Option<String>

The state of the Spot Instance request. Spot status information can help you track your Spot Instance requests. For more information, see Spot Status in the Amazon Elastic Compute Cloud User Guide.

status: Option<SpotInstanceStatus>

The status code and status message describing the Spot Instance request.

tags: Option<Vec<Tag>>

Any tags assigned to the resource.

type_: Option<String>

The Spot Instance request type.

valid_from: Option<String>

The start date of the request, in UTC format (for example, YYYY-MM-DDTHH:MM:SSZ). The request becomes active at this date and time.

valid_until: Option<String>

The end date of the request, in UTC format (for example, YYYY-MM-DDTHH:MM:SSZ). If this is a one-time request, it remains active until all instances launch, the request is canceled, or this date is reached. If the request is persistent, it remains active until it is canceled or this date is reached. The default end date is 7 days from the current date.